% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_METData_to_predict.R
\name{add_METData_to_predict}
\alias{add_METData_to_predict}
\alias{new_add_METData_to_predict}
\alias{validate_add_METData_to_predict}
\title{Create a multi-environment trials data object, which will be used for
prediction of unobserved data.}
\usage{
new_add_METData_to_predict(
  METData_training,
  pheno_new = NULL,
  info_environments_to_predict = NULL,
  geno_new = NULL,
  compute_climatic_ECs = FALSE,
  env_data_manual = NULL,
  raw_weather_data = NULL,
  ...
)

add_METData_to_predict(
  METData_training,
  pheno_new = NULL,
  geno_new = NULL,
  compute_climatic_ECs = FALSE,
  info_environments_to_predict = NULL,
  env_data_manual = NULL,
  raw_weather_data = NULL,
  ...
)

validate_add_METData_to_predict(x, ...)
}
\arguments{
\item{METData_training}{\code{METData} object, which will be used as training
set. Normally, it should be the same \code{METData} object as the one
created with \code{\link[=create_METData]{create_METData()}} and evaluated with \code{\link[=predict_trait_MET_cv]{predict_trait_MET_cv()}}.}

\item{pheno_new}{\code{data.frame} object with at least 3 columns.
\enumerate{
\item geno_ID \code{character} contains the genotype identifiers.
\item year \code{numeric} contains the year of the observation.
\item location \code{character} contains the name of the location.
}
Basically, this \code{data.frame} indicates which genotypes should be
predicted in which environments. No columns corresponding to phentoypic
traits is expected.}

\item{info_environments_to_predict}{\code{data.frame} object with at least
the 4 following columns. \cr
\enumerate{
\item year: \code{numeric} Year label of the environment
\item location: \code{character} Name of the location
\item longitude: \code{numeric} longitude of the environment
\item latitude: \code{numeric} latitude of the environment

}
The two next columns are required only if weather data should be
retrieved from NASA POWER data.
\enumerate{
\item planting.date: (optional) \code{Date} YYYY-MM-DD
\item harvest.date: (optional) \code{Date} YYYY-MM-DD \cr
}
\itemize{
\item \strong{The data.frame should contain as many rows as Year x Location
combinations which will be used in pheno_new.}
}}

\item{geno_new}{\code{data.frame} object which should contain genotypic data
for new candidates, if these are not included in \code{METData_training$geno}
which will be used as training set. If already included, no need to include
geno_new. Default is \code{NULL}.}

\item{compute_climatic_ECs}{\code{logical} indicates if environmental
covariates should be computed with the function. Default
is \code{FALSE}. \cr
\strong{Set compute_climatic_ECs = \code{TRUE} if user wants to use weather data
from NASA POWER data. For instance, if no weather-based covariables
can be provided or if raw weather data are only available for some
environments but not for others.}}

\item{env_data_manual}{\code{data.frame} can be let as NULL by user, if no
environmental covariables provided as input. Otherwise, a \code{data.frame}
should be provided.
\strong{The data.frame should contain as many rows as the phenotypic
dataset \code{data.frame}.} \cr
Columns should be:
\enumerate{
\item geno_ID \code{character} with genotype ID
\item year \code{numeric} with the year label
\item location \code{character} with the location character
}
Columns 4 and + should be numeric and contain the environmental covariates.
\cr
\itemize{
\item \strong{Providing env_data  and setting \code{compute_climatic_ECs = T} is
possible. For instance, the user can have some information regarding the
soil composition (\\% \% clay, sand, silt, organic matter content).
A disease status can also be encoded as categorical variable if it affects
some environments. In addition to these type of covariates, weather-based
covariates will be computed if \code{compute_climatic_ECs = T}.}
}}

\item{raw_weather_data}{\code{data.frame} can be let as NULL by user, if no
daily weather datasets are available. If else, required columns should be
provided like this (colnames should be respected):
\enumerate{
\item longitude \code{numeric}
\item latitude \code{numeric}
\item year \code{numeric}
\item location \code{character}
\item YYYYMMDD \code{Date}
}
Available weather data provided by user must be a subset of the following
weather variable names. Colnames must be given as following:
\enumerate{
\item T2M \code{numeric} Daily mean temperature (째C)
\item T2M_MIN \code{numeric} Daily minimum temperature (째C)
\item T2M_MAX \code{numeric} Daily maximum temperature (째C)
\item PRECTOT \code{numeric} Daily total precipitation (mm)
\item RH2M \code{numeric} Daily mean relative humidity (\%)
\item RH2M_MIN \code{numeric} Daily minimum relative humidity (\%)
\item RH2M_MAX \code{numeric} Daily maximum relative humidity (\%)
\item daily_solar_radiation \code{numeric} daily solar radiation
(MJ/m^2/day)
\item top_atmosphere_insolation \code{numeric} Top-of-atmosphere
Insolation (MJ/m^2/day)
\item T2MDEW \code{numeric} Dew Point (째C)
}

\strong{It is not required that weather data for ALL environments are
provided by the user. If weather data for some environments are missing,
they will be retrieved by the NASA }}
}
\value{
A formatted \code{list} of class \code{METData} which contains the
following elements:
\itemize{
\item \strong{geno}: \code{matrix} with genotype values of phenotyped individuals.
\item \strong{map}: \code{data.frame} with genetic map.
\item \strong{pheno}: \code{data.frame} with phenotypic trait values.
\item \strong{compute_climatic_ECs}: \code{logical} indicates if environmental
covariates were required to be retrieved via the package by the user.
\item \strong{env_data}: \code{data.frame} with the environmental covariates per
environment.
\item \strong{info_environments}: \code{data.frame} contains basic information on
each environment.
\item \strong{ECs_computed}: OPTIONAL \code{logical} subelement added in the output
only if the function \code{\link[=get_ECs]{get_ECs()}} was correctly run within the pipeline.
}
}
\description{
This function combines all types of predictor sources (genotypic,
information about the environments to be predicted, environmental data if
available...) and the data.frame describing the phenotypic observations to
be predicted.
}
\examples{

}
\author{
Cathy C. Jubin \email{cathy.jubin@uni-goettingen.de}
}
