% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_new_METData.R
\name{add_new_METData}
\alias{add_new_METData}
\title{Create a multi-environment trials data object}
\usage{
add_new_METData(
  METData_training,
  geno_new = NULL,
  pheno_new,
  info_environments_new = NULL,
  unique_EC_by_geno = FALSE,
  scenario_weather_data,
  env_data_manual = NULL,
  crop = NULL
)
}
\arguments{
\item{METData_training}{Will be used as training set.}

\item{pheno_new}{\code{data.frame} object with at least 3 columns.
\enumerate{
\item geno_ID \code{character} contains the genotype identifiers.
\item year \code{numeric} contains the year of the observation.
\item location \code{character} contains the name of the location.
}
Basically, this \code{data.frame} only indicates which genotypes should be
predicted in which environments. No columns with phenotypic values is
expected.}

\item{unique_EC_by_geno}{\code{logical} indicates if the environmental
covariates contained in env_data_manual are also genotype-specific (dependent on
the phenology, for instance) or unique for a complete environment. Default
is \code{FALSE}.}

\item{scenario_weather_data}{\code{character}. Options are:
\itemize{
\item \strong{\code{manually}}: names of columns in the \code{env_data_manual} data.frame should be the
same as in original METData$env_data).
\item \strong{\code{use_real_EC}}: environmental variables will be estimated for a set of "past" environments based on
information provided in info_environments_new.
using the \code{\link[=use_real_EC]{use_real_EC()}} function. This assumes that the environmental data can be
retrieved and are from the past. only possible if...
different according to compute_EC.
\item \strong{\code{mean_previous_years}}: ECs computed based on average over 5 last years
of the location indicated in info_environments_new
use of the location data from the location in the training set (implies
that ) if computed_ECs== TRUE in METData_training.
give fictive harvest date --> example
\item \strong{\code{previous_year}}:
}}

\item{env_data_manual}{\code{data.frame}
can contain soil variables (with same column names as used in METData_training$env_data) but no weather-based covariates

can be let as NULL by user, if no
environment data provided as input. Otherwise, a \code{data.frame} should
be provided:
Two types of \code{data.frame} can be provided:
\itemize{
\item \strong{Type 1}: if \code{unique_EC_by_geno = FALSE}. Each environmental covariate
characterizes all phenotypes within a complete environment (e.g. ECs are
not computed individually for each variety within an environment).
\strong{The data.frame should contain as many rows as
the info_environments \code{data.frame}.} \cr
Columns should contain:
\enumerate{
\item year \code{numeric} with the year label
\item location \code{character} with the location character
}
Columns 3 and + should be numeric and contain the environmental covariates.
\cr
}

OR \cr
\itemize{
\item \strong{Type 2}: if \code{unique_EC_by_geno = TRUE}. Each environmental covariate
is computed specifically for an environment AND for a genotype (e.g. ECs
are computed individually for each variety within an environment).
\strong{The data.frame should contain as many rows as the phenotypic
dataset \code{data.frame}.} \cr
Columns should contain:
\enumerate{
\item geno_ID \code{character} with genotype ID
\item year \code{numeric} with the year label
\item location \code{character} with the location character
}
Columns 4 and + should be numeric and contain the environmental covariates.
\cr
}}

\item{info_environments}{\code{data.frame} object with at least the 4 first
columns. \cr
\enumerate{
\item year: \code{numeric} Year label of the environment
\item location: \code{character} Name of the location
\item longitude: \code{numeric} longitude of the environment
\item latitude: \code{numeric} latitude of the environment
\item planting.date: (optional) \code{Date} YYYY-MM-DD
\item harvest.date: (optional) \code{Date} YYYY-MM-DD \cr
}
\itemize{
\item \strong{The data.frame should contain as many rows as Year x Location
combinations. Example: if only one location evaluated across four years, 4
rows should be present.}
\item \strong{The fifth and sixth columns (planting.date and harvest.date) are
required only if the user wants to download weather data with the
package (setting argument \code{ECs_computed = T} in \code{\link[=create_METData]{create_METData()}} and using
subsequently the function \code{\link[=get_ECs]{get_ECs()}}).}
}}

\item{location_to_use}{\code{character}. Default is NULL.}

\item{filtering_markers}{\code{logical} indicator if a single-environment
QTL detection step (performed with GWAS or ElasticNet using the function
\code{\link[=select_markers]{select_markers()}}) should be conducted, to identify a subset of markers
representing potential environment-specific QTL effects, for which
interactions with environmental covariates can be further investigated.
Default is \code{TRUE}.}
}
\value{
A formatted \code{list} of class \code{METData} which contains the
following elements:
\itemize{
\item \strong{geno}: \code{matrix} with genotype values of phenotyped individuals.
\item \strong{map}: \code{data.frame} with genetic map.
\item \strong{pheno}: \code{data.frame} with phenotypic trait values.
\item \strong{compute_EC_by_geno}: \code{logical} indicates if environmental
covariates should be later computed.
\item \strong{env_data}: \code{data.frame} with the environmental covariates per
environment (and if genotype-specific, per genotype).
\item \strong{info_environments}: \code{data.frame} contains basic information on
each environment.
\item \strong{unique_EC_by_geno}: \code{logical} to indicate if the EC is genotype-
specific.
\item \strong{filtering_markers}: \code{logical} indicates if a filtering marker
step should be applied in further steps
}
}
\description{
This function combines all types of data sources (genotypic, phenotypic,
environmental data, information about the environments...) in a single data
object of class \code{METData}.
}
\examples{

data(geno_japonica)
data(map_japonica)
data(pheno_japonica)
pheno_japonica2<-pheno_japonica[which(pheno_japonica$year==2013),]
data(info_environments_japonica)
info_environments_japonica2<-info_environments_japonica[which(info_environments_japonica$year==2013),]
data(env_data_japonica)
env_data_japonica2<-env_data_japonica[which(env_data_japonica$year==2013),]


}
\author{
Cathy C. Jubin \email{cathy.jubin@uni-goettingen.de}
}
