% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ale_plots.R
\name{ALE_plot_split}
\alias{ALE_plot_split}
\title{ALE plots feature-wise}
\usage{
ALE_plot_split(object, variable, path_plot, nb_bins = 10)
}
\arguments{
\item{object}{an object of class \code{res_fitted_split}}

\item{variable}{environmental or genomic variable (PC derived from SNPs
matrix or single SNP) for which the ALE plot should be calculated}

\item{path_plot}{path where the ALE plot should be saved}

\item{nb_bins}{number of intervals. Note that if \code{nb_bins} is provided by the
user, the \code{min_interval_points} argument is ignored.}

\item{min_interval_points}{minimum number of points contained within each
interval (e.g. among the unique values taken by the feature). Values less
than 4 should be avoided.}
}
\description{
Makes an ALE plot for one single predictor variable of the model.
}
\author{
Cathy C. Westhues \email{cathy.jubin@uni-goettingen.de}
}
