% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_predictions_best_genotypes.R
\name{analysis_predictions}
\alias{analysis_predictions}
\title{Analysis of prediction results from \code{\link[=predict_trait_MET]{predict_trait_MET()}} by location.}
\usage{
analysis_predictions(met_pred, path_save_results, trait = NULL)
}
\arguments{
\item{met_pred}{Output from \code{\link[=predict_trait_MET]{predict_trait_MET()}} function of class
\code{met_pred}.}

\item{path_save_results}{Path where plots and results should be saved.}
}
\description{
This function analyses at each individual location the best genotypes based
on their predicted performance over years. List of mean genotype
performance (simple average based on predictions) at each location are saved.

No output from this function, all results are direcly saved based on the
\code{path_save_results} argument.
}
\author{
Cathy C. Westhues \email{cathy.jubin@uni-goettingen.de}
}
