% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_predictions_best_genotypes.R
\name{analysis_predictions_best_genotypes}
\alias{analysis_predictions_best_genotypes}
\title{Analysis of prediction results from \code{\link[=predict_trait_MET]{predict_trait_MET()}} by location and by
environmental cluster.}
\usage{
analysis_predictions_best_genotypes(
  met_pred,
  path_save_results,
  cluster_envs = TRUE,
  env_predictors_for_clustering = NULL,
  K = 3
)
}
\arguments{
\item{met_pred}{Output from \code{\link[=predict_trait_MET]{predict_trait_MET()}} function of class
\code{met_pred}.}

\item{path_save_results}{Path where plots and results should be saved.}

\item{cluster_envs}{\code{logical} indicates if k-means algorithm
on environmental data should be done. Default is \code{TRUE}.}

\item{env_predictors_for_clustering}{\code{character} Vector giving the
names of environmental covariates to be used in the K-means clustering
analysis.}

\item{K}{\code{integer} number of clusters to use in K-means algorithm}
}
\description{
This function analyses at each individual location the best genotypes based
on their predicted performance over multiple years. List of mean genotype
performance (based on predictions) at each location are saved.
Second part of the function performs a K-means clustering of environments
based on the env. covariables characterizing each of these environments.
List of mean genotype performance (based on predictions) for each
environmental cluster are saved.
No output from this function, all results are direcly saved based on the
\code{path_save_results} argument.
}
\author{
Cathy C. Westhues \email{cathy.jubin@uni-goettingen.de}
}
