% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_pcs_G.R
\name{apply_pcs_G}
\alias{apply_pcs_G}
\title{Data dimensionality reduction by modeling genetic effects using the PCs of
the genomic relationship matrix.}
\usage{
apply_pcs_G(split, geno_data, num_pcs = 200, ...)
}
\arguments{
\item{split}{An object of class \code{split}, corresponding to one element of
the total \code{cv_object} generated by one of the functions \code{\link[=predict_cv0]{predict_cv0()}},
\code{\link[=predict_cv1]{predict_cv1()}}, or \code{\link[=predict_cv2]{predict_cv2()}}, and containing the following items:
\itemize{
\item \strong{training}: \code{data.frame} Training dataset
\item \strong{test}: \code{data.frame} Test dataset
}}

\item{geno_data}{\code{data.frame} An element named \code{geno} within an object of
class \code{METData}.}
}
\value{
pc_values A \code{data.frame} containing the principal components
in columns and the names of all lines used in the study is contained in the
first column 'geno_ID'. PCs for the lines present in the test set were
computed based on the transformation done on the training set.
}
\description{

}
\author{
Cathy C. Jubin \email{cathy.jubin@uni-goettingen.de}
}
