% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_pcs_G_Add.R
\name{apply_pcs_G_Add}
\alias{apply_pcs_G_Add}
\title{Data dimensionality reduction by modeling genetic effects using the PCs of
the genomic relationship matrix.}
\usage{
apply_pcs_G_Add(split, geno, num_pcs = 200, ...)
}
\arguments{
\item{split}{An object of class \code{split}, corresponding to one element of
the total \code{cv_object} generated by one of the functions \code{\link[=predict_cv0]{predict_cv0()}},
\code{\link[=predict_cv00]{predict_cv00()}}, \code{\link[=predict_cv1]{predict_cv1()}}, or \code{\link[=predict_cv2]{predict_cv2()}},
and containing the following items:
\itemize{
\item \strong{training}: \code{data.frame} Training dataset
\item \strong{test}: \code{data.frame} Test dataset
}}

\item{geno}{\code{data.frame} It corresponds to a \code{geno} element
within an object of class \code{METData}.}

\item{num_pcs}{\code{integer} Maximal number of principal components to
extract.}
}
\value{
pc_values A \code{data.frame} containing the principal components
in columns and the names of all lines used in the study is contained in the
first column 'geno_ID'. PCs for the lines present in the test set were
computed based on the transformation done on the training set.
}
\description{
Computes the genomic relationship after centering and scaling genotype matrix
of the training set. Eigenvalues and eigenvectors of the G matrix are
estimated. PCA is applied on the training set and the same transformation is
applied on the test set. The goal is to use principal components in
prediction models as a smaller number of variables instead of all the marker
predictors.
}
\references{
\insertRef{de2020data}{learnMET}
}
\author{
Cathy C. Westhues \email{cathy.jubin@uni-goettingen.de}
}
