% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\docType{data}
\name{climate_variables_japonica}
\alias{climate_variables_japonica}
\title{Multi-year trial data of rice}
\format{
\code{data.frame} containing as many rows as info_environments
Columns are:
\enumerate{
\item geno_ID \code{character} with genotype ID
\item year \code{numeric} with the year label
\item location \code{character} with the location character
\item Col 3 to Col 56: climatic covariates derived for each crop
growing stage. More explanation on the ECs in the original publication
from Monteverde et al. (2019)
}
}
\source{
Data from the INIA's Rice Breeding Program (Uruguay). Data downloaded from publication Monteverde et al. (2019).
}
\description{
This data are used as toy data for several functions of
\pkg{learnMET} package.\cr
Climate variables data for the \emph{japonica} population grown in one location (Treinta y Tres, Uruguay) for three years.
}
\references{
\insertRef{monteverde2019integrating}{learnMET}
Instituto Nacional de Investigación Agropecuaria (INIA). Programa Arroz. Estación Experimental INIA Treinta y Tres, Uruguay
}
\author{
Cathy C. Westhues \email{cathy.jubin@uni-goettingen.de}
}
\keyword{data}
