% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_env_data.R
\name{clustering_env_data}
\alias{clustering_env_data}
\title{Clustering of environments solely based on environmental information}
\usage{
clustering_env_data(weather_ECs = NULL, soil_ECs = NULL, path_plots = NULL)
}
\arguments{
\item{weather_ECs}{a \code{data.frame} with climate-based covariates (predictor variables)
with three columns year, location and IDenv.
Default can be \code{NULL}, if no weather-based data available/should be used.}

\item{soil_ECs}{a \code{data.frame} with soil-based covariates (predictor variables)
with three columns year, location and IDenv.
Default can be \code{NULL}, if no soil data available.}

\item{path_plots}{\code{character} Path where clustering plots should be saved.}

\item{k}{\code{numeric} Number of clusters to use.}
}
\description{
Clustering of environments using a K-means algorithm based on:
\enumerate{
\item Only climate variables
\item Only soil variables
\item All environmental data together
}
}
\author{
Cathy C. Westhues \email{cathy.jubin@uni-goettingen.de}
}
