% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_EC_gdd.R
\name{compute_EC_gdd}
\alias{compute_EC_gdd}
\title{Compute ECs based on growth requirements estimated from accumulated GDD.}
\usage{
compute_EC_gdd(
  table_daily_W,
  crop_model,
  method_GDD_calculation = c("method_b"),
  ...
)
}
\description{
This function enables to retrieve daily weather data for each
environment and derive environmental covariates over non-overlapping time
windows, which can be defined in various ways by the user.
}
\author{
Cathy C. Jubin \email{cathy.jubin@uni-goettingen.de}
}
