% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_EC_user_defined_intervals.R
\name{compute_EC_user_defined_intervals}
\alias{compute_EC_user_defined_intervals}
\title{Compute ECs based on day-windows of fixed length.}
\usage{
compute_EC_user_defined_intervals(
  table_daily_W,
  intervals_growth_manual = NULL,
  base_temperature = 10,
  capped_max_temperature = F,
  max_temperature = 35,
  method_GDD_calculation = c("method_b"),
  ...
)
}
\arguments{
\item{table_daily_W}{\code{data.frame} returned by the function
\code{\link[=get_daily_tables_per_env]{get_daily_tables_per_env()}}}

\item{intervals_growth_manual}{\code{data.frame} with:
\itemize{
\item column 1: \code{numeric} year
\item column 2: \code{character} location
\item columns 3 and +: \code{numeric} Date (in Days after Planting) at which
the crop enters a new growth stage in a given environment.
"P" refers to the planting date and should contain 0 as value, "VE" to
emergence, etc...
\strong{Day 0 (Planting Date, denoted "P") should be in the third column.
At least 4 columns should be in this data.frame. There is no need to
indicate the column "Harvest" - already considered in the function.}
An example of how this data.frame should be provided is given in
\link{intervals_growth_manual_G2F}.\cr
}}

\item{base_temperature}{\code{numeric} Base temperature (crop growth assumed
to be null below this value.) Default is 10.}

\item{method_GDD_calculation}{\code{character} Method used to compute the
GDD value, with one out of \code{method_a} or \code{method_b}. \cr
\code{method_a}: No change of the value of \eqn{T_{min}}.
GDD = \eqn{max (\frac{T_{min}+T_{max}}{2} - T_{base},0)}. \cr
\code{method_b}: If \eqn{T_{min}} < \eqn{T_{base}}, change \eqn{T_{min}}
to \eqn{T_{min}} = \eqn{T_{base}}. \cr
Default = \code{method_b}.}
}
\value{
An object of class \code{data.frame} with
10 x number_total_fixed_windows + 1 last column (IDenv):
\enumerate{
\item mean_TMIN: number_total_fixed_windows columns, indicating the
average minimal temperature over the respective day-window.
\item mean_TMAX: number_total_fixed_windows columns, indicating the
average maximal temperature over the respective day-window.
\item mean_TMEAN: number_total_fixed_windows columns, indicating the
average mean temperature over the respective day-window.
\item freq_TMAX_sup30: number_total_fixed_windows columns, indicating the
frequency of days with maximum temperature over 30°C over the respective
day-window.
\item freq_TMAX_sup35: number_total_fixed_windows columns, indicating the
frequency of days with maximum temperature over 35°C over the respective
day-window.
\item sum_GDD: number_total_fixed_windows columns, indicating the
growing degree days over the respective day-window.
\item sum_PTT: number_total_fixed_windows columns, indicating the
accumulated photothermal time over the respective day-window.
\item sum_P: number_total_fixed_windows columns, indicating the
accumulated precipitation over the respective day-window.
\item sum_et0: number_total_fixed_windows columns, indicating the
cumulative reference evapotranspiration over the respective day-window.
\item freq_P_sup10: number_total_fixed_windows columns, indicating the
frequency of days with total precipitation superior to 10 mm over the
respective day-window.
\item sum_solar_radiation: number_total_fixed_windows columns, indicating
the accumulated incoming solar radiation over the respective day-window.
\item mean_vapr_deficit: number_total_fixed_windows columns, indicating
the mean vapour pressure deficit over the respective day-window.
\item IDenv \code{character} ID of the environment (Location_Year)
}
}
\description{
Compute the environmental covariates based on the daily weather
table of an environment (Year x Location), and over day-windows which can be
defined by the user for each environment (based for instance on observed
phenological dates) in a table provided as input in \code{\link[=create_METData]{create_METData()}}.
}
\author{
Cathy C. Westhues \email{cathy.jubin@uni-goettingen.de}
}
