% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_EC.R
\name{daylength}
\alias{daylength}
\title{Compute the day length given the altitude and day of year.}
\usage{
daylength(lat, day_of_year)
}
\arguments{
\item{day_of_year}{\code{numeric} Day of year}

\item{latitude}{\code{numeric} Latitude}
}
\value{
\code{numeric} Number of hours of daytime.
}
\description{
Compute the day length given the altitude and day of year.
}
\references{
A model comparison for daylength as a function of latitude and
day of year. Ecological Modelling, 80(1), 87-95. Forsythe, W. C., Rykiel Jr
,E. J., Stahl, R. S., Wu, H. I., & Schoolfield, R. M. (1995).
}
\author{
Cathy C. Westhues \email{cathy.jubin@hotmail.com}
}
