% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DL_reg.R
\name{DL_reg}
\alias{DL_reg}
\alias{new_DL_reg}
\alias{validate_DL_reg}
\title{deep learning reg}
\usage{
new_DL_reg(
  split,
  trait,
  geno_data,
  env_predictors,
  info_environments,
  unique_EC_by_geno,
  geno_information,
  use_selected_markers,
  SNPs,
  list_env_predictors,
  include_env_predictors,
  lat_lon_included,
  year_included,
  ...
)

DL_reg(
  split,
  trait,
  geno_data,
  env_predictors,
  info_environments,
  unique_EC_by_geno,
  geno_information,
  use_selected_markers,
  SNPs,
  list_env_predictors,
  include_env_predictors,
  lat_lon_included,
  year_included,
  ...
)

validate_DL_reg(x, ...)
}
\description{
deep learning reg
}
