% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtw_distance.R
\name{dtw_distance}
\alias{dtw_distance}
\title{Compute the DTW distance using the dtwcluster R package & compute features}
\usage{
dtw_distance(daily_weather_data, ...)
}
\arguments{
\item{a}{list of data.frame with the following columns extracted
from NASA POWER data, according to requested parameters:
\enumerate{
\item longitude \code{numeric}
\item latitude \code{numeric}
\item YEAR \code{numeric}
\item MM \code{integer}
\item DD \code{integer}
\item DOY \code{integer}
\item YYYYMMDD \code{Date}
\item RH2M \code{numeric}
\item T2M \code{numeric}
\item T2M_MIN \code{numeric}
\item T2M_MAX \code{numeric}
\item PRECTOTCORR \code{numeric}
\item ALLSKY_SFC_SW_DWN \code{numeric}
\item T2MDEW \code{numeric}
\item IDenv \code{character} ID environment for which weather data were
downloaded.
\item length.gs \code{difftime} length in days of the growing season
for the environment.
}}
}
\value{
a data.frame \code{data.frame} with the results from the multi
dimensional scaling analysis.
}
\description{
Compute DTW distance using the dtw_basic function from dtwcluster R package
and  multi-dimensional scaling on the obtained distance to get features for
ML analyses
}
\author{
Cathy C. Westhues \email{cathy.jubin@hotmail.com}
}
