% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\docType{data}
\name{env_data_indica}
\alias{env_data_indica}
\title{Multi-year trial data of rice}
\format{
\code{data.frame} containing as many rows as info_environments
Columns are:
\enumerate{
\item geno_ID \code{character} with genotype ID
\item year \code{numeric} with the year label
\item location \code{character} with the location character
\item Col 3 to Col 56: environmental covariates derived for each crop
growing stage. More explanation on the ECs in the original publication
from Monteverde et al. (2019)
}
}
\source{
Data from the INIA's Rice Breeding Program (Uruguay). Data downloaded from publication Monteverde et al. (2019).
}
\description{
This data are used as toy data for several functions of
\pkg{learnMET} package.
}
\references{
\insertRef{monteverde2019integrating}{learnMET}
}
\author{
Cathy C. Westhues \email{cathy.jubin@uni-goettingen.de}
}
\keyword{data}
