% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_predict_model.R
\name{fit_predict_model}
\alias{fit_predict_model}
\alias{fit_predict_model.default}
\alias{fit_predict_model.xgb_reg}
\alias{fit_predict_model.xgb_ordinal}
\alias{fit_predict_model.DL_reg}
\alias{fit_predict_model.svm_stacking_reg}
\title{fit_predict_model}
\usage{
fit_predict_model(object, ...)

\method{fit_predict_model}{default}(x, ...)

\method{fit_predict_model}{xgb_reg}(object, seed, inner_cv_reps = 1, inner_cv_folds = 3, ...)

\method{fit_predict_model}{xgb_ordinal}(object, seed, inner_cv_reps = 2, inner_cv_folds = 5, ...)

\method{fit_predict_model}{DL_reg}(object, seed, inner_cv_reps = 1, inner_cv_folds = 3, ...)

\method{fit_predict_model}{svm_stacking_reg}(
  object,
  seed,
  inner_cv_reps = 2,
  inner_cv_folds = 5,
  kernel_G = "rbf",
  kernel_E = "rbf",
  kernel_GE = "rbf",
  ...
)
}
\arguments{
\item{object}{It must be an object of class}

\item{fit}{CV object}
}
\description{
for all CV schemes, the prediction accuracy is computed as the correlations between the observed
and predicted values within same environments.
}
\author{
Cathy C. Jubin \email{cathy.jubin@uni-goettingen.de}
}
