% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_DL_reg_1.R, R/fit_DL_reg_2.R,
%   R/fit_DL_reg_3.R, R/fit_rf_reg_1.R, R/fit_rf_reg_2.R, R/fit_rf_reg_3.R,
%   R/fit_split.R, R/fit_stacking_reg_1.R, R/fit_stacking_reg_2.R,
%   R/fit_stacking_reg_3.R, R/fit_xgb_reg_1.R, R/fit_xgb_reg_2.R,
%   R/fit_xgb_reg_3.R
\name{fit_split.DL_reg_1}
\alias{fit_split.DL_reg_1}
\alias{fit_split.DL_reg_2}
\alias{fit_split.DL_reg_3}
\alias{fit_split.rf_reg_1}
\alias{fit_split.rf_reg_2}
\alias{fit_split.rf_reg_3}
\alias{fit_split}
\alias{fit_split.default}
\alias{fit_split.stacking_reg_1}
\alias{fit_split.stacking_reg_2}
\alias{fit_split.stacking_reg_3}
\alias{fit_split.xgb_reg_1}
\alias{fit_split.xgb_reg_2}
\alias{fit_split.xgb_reg_3}
\title{S3 method used to fit an object of class  \code{rf_reg_1}, \code{rf_reg_2},
\code{rf_reg_3}, \code{xgb_reg_1}, \code{xgb_reg_2}, \code{xgb_reg_3},\code{DL_reg},\code{DL_reg_1},
\code{DL_reg_2},\code{DL_reg_3},\code{stacking_reg_1}, \code{stacking_reg_2} or \code{stacking_reg_3}.}
\usage{
\method{fit_split}{DL_reg_1}(
  object,
  seed,
  inner_cv_reps = 1,
  inner_cv_folds = 3,
  save_model = F,
  ...
)

\method{fit_split}{DL_reg_2}(
  object,
  seed,
  inner_cv_reps = 1,
  inner_cv_folds = 3,
  save_model = F,
  ...
)

\method{fit_split}{DL_reg_3}(
  object,
  seed,
  inner_cv_reps = 1,
  inner_cv_folds = 3,
  save_model = F,
  ...
)

\method{fit_split}{rf_reg_1}(
  object,
  seed,
  inner_cv_reps = 1,
  inner_cv_folds = 5,
  save_model = F,
  ...
)

\method{fit_split}{rf_reg_2}(
  object,
  seed,
  inner_cv_reps = 1,
  inner_cv_folds = 5,
  save_model = F,
  ...
)

\method{fit_split}{rf_reg_3}(
  object,
  seed,
  inner_cv_reps = 1,
  inner_cv_folds = 5,
  save_model = F,
  ...
)

fit_split(object, ...)

\method{fit_split}{default}(x, ...)

\method{fit_split}{stacking_reg_1}(
  object,
  seed,
  inner_cv_reps = 1,
  inner_cv_folds = 5,
  kernel_G = "linear",
  kernel_E = "polynomial",
  path_folder,
  save_model = F,
  ...
)

\method{fit_split}{stacking_reg_2}(
  object,
  seed,
  inner_cv_reps = 1,
  inner_cv_folds = 5,
  kernel_G = "linear",
  kernel_E = "polynomial",
  kernel_GE = "polynomial",
  save_model = F,
  ...
)

\method{fit_split}{stacking_reg_3}(
  object,
  seed,
  inner_cv_reps = 1,
  inner_cv_folds = 5,
  kernel_G = "linear",
  kernel_E = "polynomial",
  save_model = F,
  ...
)

\method{fit_split}{xgb_reg_1}(
  object,
  seed,
  inner_cv_reps = 1,
  inner_cv_folds = 2,
  save_model = F,
  ...
)

\method{fit_split}{xgb_reg_2}(
  object,
  seed,
  inner_cv_reps = 1,
  inner_cv_folds = 5,
  save_model = F,
  ...
)

\method{fit_split}{xgb_reg_3}(
  object,
  seed,
  inner_cv_reps = 1,
  inner_cv_folds = 5,
  save_model = F,
  ...
)
}
\arguments{
\item{object}{an object of class \code{rf_reg_1}, \code{rf_reg_2}, \code{rf_reg_3},
\code{xgb_reg_1}, \code{xgb_reg_2}, \code{xgb_reg_3},\code{DL_reg},\code{DL_reg_1},\code{DL_reg_2},
\code{DL_reg_3},\code{stacking_reg_1}, \code{stacking_reg_2} or \code{stacking_reg_3}.}
}
\description{
S3 dispatching method for objects of class \code{rf_reg_1}, \code{rf_reg_2},
\code{rf_reg_3}, \code{xgb_reg_1}, \code{xgb_reg_2}, \code{xgb_reg_3},\code{DL_reg},\code{DL_reg_1},
\code{DL_reg_2},\code{DL_reg_3},\code{stacking_reg_1}, \code{stacking_reg_2} or \code{stacking_reg_3}.
}
\author{
Cathy C. Westhues \email{cathy.jubin@uni-goettingen.de}
}
