% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_train_test_split.R
\name{fitting_train_test_split}
\alias{fitting_train_test_split}
\title{Processing and selecting predictors to fit the model}
\usage{
fitting_train_test_split(split, prediction_method)
}
\description{
Processing and selecting predictors to fit the model
}
