% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdd_information.R
\name{gdd_information}
\alias{gdd_information}
\title{Internal function of [compute_EC_gdd())]

[compute_EC_gdd())]: R:compute_EC_gdd())}
\usage{
gdd_information(crop_model)
}
\arguments{
\item{crop_model}{name of the crop model to be used to estimate growth stage
based on growing degree days accumulation. \cr
Current possible values are: 'maizehybrid1700' and 'hardwheatUS',}
}
\value{
a \code{data.frame} with:
\enumerate{
\item Stage: \code{character} NAme of the stage
\item GDD: \code{numeric} Accumulated GDDs to reach this growth stage
}
}
\description{
Get the table of estimated growth stage based on GDD (in Celsius degrees)
Only maize (Zea mays) and hard wheat (Triticum durum) currently implemented.
}
\author{
Cathy C. Westhues \email{cathy.jubin@hotmail.com}
}
