% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\docType{data}
\name{geno_G2F}
\alias{geno_G2F}
\title{Maize experimental multi-environment data sets (Genomes to Fields Initiative)}
\format{
\code{data.frame} with genotype values of maize hybrids (\emph{in silico} genotype matrix based on inbred lines)
stored, with geno_ID as row.names and markers in columns.
Dimensions are \code{n}=1,913 and \code{p}=106,414.
}
\source{
Subset of the publicly available datasets from the Genomes to Fields Initiative. \cr
All data are publicly available at: https://www.genomes2fields.org/resources/,
}
\description{
This data are used as toy data for several functions of
\pkg{learnMET} package. \cr
Physical map (SNP information generated using a genotyping-by-sequence (GBS) method for the inbred lines) for
106,414 markers.
}
\references{
\insertRef{mcfarland2020maize}{learnMET} \
\insertRef{alkhalifah2018maize}{learnMET}
}
\author{
Cathy C. Westhues \email{cathy.jubin@uni-goettingen.de}
}
\keyword{data}
