% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapor_pressure.R
\name{get.ea.no.RH}
\alias{get.ea.no.RH}
\title{Formulas to compute vapour pressure deficit according to available data}
\usage{
get.ea.no.RH(tmin)
}
\arguments{
\item{tmin}{\code{numeric} minimum daily air temperature, °C}
}
\value{
Actual vapor pressure (ea)
}
\description{
Actual vapor pressure (ea) derived from relative humidity calculated
with  tmin if no data related to humidity are available.
}
\references{
\insertRef{zotarelli2010step}{learnMET}
}
\author{
Cathy C. Westhues \email{cathy.jubin@hotmail.com}
}
