% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapor_pressure.R
\name{get.ea}
\alias{get.ea}
\title{Formulas to compute vapour pressure deficit according to available data}
\usage{
get.ea(rhmin, rhmax, tmin, tmax)
}
\arguments{
\item{rhmin}{\code{numeric} minimum daily relative humidity, \%}

\item{rhmax}{\code{numeric} maximum daily relative humidity, \%}

\item{tmin}{\code{numeric} minimum daily air temperature, °C}

\item{tmax}{\code{numeric} maximum daily air temperature, °C}
}
\value{
Actual vapor pressure (ea)
}
\description{
Actual vapor pressure (ea) derived from relative humidity calculated
with rhmin and rhmax.
}
\references{
\insertRef{zotarelli2010step}{learnMET}
}
\author{
Cathy C. Westhues \email{cathy.jubin@uni-goettingen.de}
}
