% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapor_pressure.R
\name{get.ea.with.rhmax}
\alias{get.ea.with.rhmax}
\title{Formulas to compute vapour pressure deficit according to available data}
\usage{
get.ea.with.rhmax(tmin, rhmax)
}
\arguments{
\item{tmin}{\code{numeric} minimum daily air temperature, °C}

\item{rhmax}{\code{numeric} maximum daily relative humidity, \%}
}
\value{
Actual vapor pressure (ea)
}
\description{
Actual vapor pressure (ea) derived from relative humidity calculated
with only rhmax.
}
\references{
\insertRef{zotarelli2010step}{learnMET}
}
\author{
Cathy C. Westhues \email{cathy.jubin@hotmail.com}
}
