% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapor_pressure.R
\name{get.ea.with.rhmean}
\alias{get.ea.with.rhmean}
\title{Formulas to compute vapour pressure deficit according to available data}
\usage{
get.ea.with.rhmean(tmin, tmax, rhmean)
}
\arguments{
\item{tmin}{\code{numeric} minimum daily air temperature, °C}

\item{tmax}{\code{numeric} maximum daily air temperature, °C}

\item{rhmean}{\code{numeric} mean daily relative humidity, \%}
}
\value{
Actual vapor pressure (ea)
}
\description{
Actual vapor pressure (ea) derived from relative humidity calculated
with rhmean, if rhmin and rhmax not available or bad data quality.
}
\references{
\insertRef{zotarelli2010step}{learnMET}
}
\author{
Cathy C. Westhues \email{cathy.jubin@hotmail.com}
}
