% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapor_pressure.R
\name{get.es}
\alias{get.es}
\title{Formulas to compute vapour pressure deficit according to available data}
\usage{
get.es(tmin, tmax)
}
\arguments{
\item{tmin}{\code{numeric} minimum daily air temperature, °C}

\item{tmax}{\code{numeric} maximum daily air temperature, °C}
}
\value{
mean saturation vapor pressure is calculated as the mean between the
saturation vapor pressure at both the daily maximum and minimum air
temperatures.
}
\description{
Formula to calculate mean saturation vapor pressure from esmx and esmn
}
\references{
\insertRef{zotarelli2010step}{learnMET}
}
\author{
Cathy C. Westhues \email{cathy.jubin@uni-goettingen.de}
}
