% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapor_pressure.R
\name{get.esmn}
\alias{get.esmn}
\title{Formulas to compute vapour pressure deficit according to available data}
\usage{
get.esmn(tmin)
}
\arguments{
\item{tmin}{\code{numeric} minimum daily air temperature, °C}
}
\value{
saturation vapor pressure at the daily minimum air temperature
}
\description{
Formula to calculate Saturation vapor pressure (es) at the daily minimum air
temperature
}
\references{
\insertRef{zotarelli2010step}{learnMET}
}
\author{
Cathy C. Jubin \email{cathy.jubin@uni-goettingen.de}
}
