% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapor_pressure.R
\name{get.esmx}
\alias{get.esmx}
\title{Formulas to compute vapour pressure deficit according to available data}
\usage{
get.esmx(tmax)
}
\arguments{
\item{tmax}{\code{numeric} maximum daily air temperature, °C}
}
\value{
saturation vapor pressure at the daily maximum air temperature
}
\description{
Formula to calculate Saturation vapor pressure (es) at the daily maximum air
temperature
}
\references{
\insertRef{zotarelli2010step}{learnMET}
}
\author{
Cathy C. Westhues \email{cathy.jubin@uni-goettingen.de}
}
