% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_daily_tables_per_env.R
\name{get_daily_tables_per_env}
\alias{get_daily_tables_per_env}
\title{Obtain daily climate data for an environment based on longitude, latitude,
planting and harvest date characterizing this environment.}
\usage{
get_daily_tables_per_env(environment, info_environments)
}
\arguments{
\item{environment}{\code{character} Name of the environment for which climate
data should be extracted.}

\item{info_environments}{\code{data.frame} with at least 4 columns.
\enumerate{
\item year \code{numeric} Year label of the environment
\item location \code{character} Name of the location
\item longitude \code{numeric} longitude of the environment
\item latitude \code{numeric} latitude of the environment
\item IDenv \code{character} ID environment (combination Year x Location)

Additional (optional) columns, required if the user wants to download
weather data with the package (via argument compute_ECs = T in
create_METData() ):
\item planting.date \code{Date} YYYY-MM-DD
\item harvest.date \code{Date} YYYY-MM-DD
\strong{The data.frame should contain as many rows as Year x Location
combinations. Example: if only one location used in the analyses for four
years, 4 rows should be present (same information with only the value in
column year changing).}
}}
}
\value{
a data.frame \code{data.frame} with the following columns extracted
from POWER data, according to requested parameters:
\enumerate{
\item LON \code{numeric}
\item LAT \code{numeric}
\item YEAR \code{numeric}
\item MM \code{integer}
\item DD \code{integer}
\item DOY \code{integer}
\item YYYYMMDD \code{Date}
\item RH2M \code{numeric}
\item T2M \code{numeric}
\item T2M_MIN \code{numeric}
\item T2M_MAX \code{numeric}
\item PRECTOT \code{numeric}
\item ALLSKY_TOA_SW_DWN \code{numeric}
\item ALLSKY_SFC_SW_DWN \code{numeric}
\item T2MDEW \code{numeric}
\item WS2M \code{numeric}
\item IDenv \code{character} ID environment for which weather data were
downloaded.
\item length.gs \code{difftime} length in days of the growing season
for the environment.
}
}
\description{
Obtain daily climate data for an environment based on longitude, latitude,
planting and harvest date characterizing this environment.
}
