% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ECs.R
\name{get_ECs}
\alias{get_ECs}
\title{Compute environmental covariates for each environment of the MET dataset
based on geographical coordinates and a given time frame.}
\usage{
get_ECs(
  METData,
  unique_EC_by_geno = F,
  fixed_length_time_windows_across_env = T,
  duration_time_window_days = 15,
  fixed_nb_windows_across_env = F,
  nb_windows_intervals = 5,
  customized_growth_intervals = F,
  save_daily_weather_tables = T,
  path_daily_weather_tables,
  ...
)
}
\description{
This function enables to retrieve daily weather data for each
environment and derive environmental covariates over time windows, which
can be defined in various ways.
}
\author{
Cathy C. Jubin \email{cathy.jubin@uni-goettingen.de}
}
