% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elevation.R
\name{get_elevation}
\alias{get_elevation}
\title{Obtain elevation data for each field trial based on longitude and latitude
data}
\usage{
get_elevation(info_environments, path)
}
\arguments{
\item{info_environments}{\code{data.frame} object with at least the 4 first
columns. \cr
\enumerate{
\item year: \code{numeric} Year label of the environment
\item location: \code{character} Name of the location
\item longitude: \code{numeric} longitude of the environment
\item latitude: \code{numeric} latitude of the environment
\item planting.date: (optional) \code{Date} YYYY-MM-DD
\item harvest.date: (optional) \code{Date} YYYY-MM-DD
\item IDenv: \code{character} ID of the environment (location x year)\cr
}

@return elevation_data\code{data.frame} object with 2 columns. \cr
\enumerate{
\item IDenv
\item elevation
}}
}
\description{
A few steps are required to get the elevation data for a given field trial
(there might be a more straightforward manner, though):
\enumerate{
\item Get the country name based on the map.where function from the maps
package
\item Convert the country name to ISO code for countries
\item Use the getData function from raster package using the ISO codes
and the longitude and latitude data to obtain elevation
}
}
