% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_soil_data.R
\name{get_soil_per_env}
\alias{get_soil_per_env}
\title{Obtain soil data for a given environment}
\usage{
get_soil_per_env(environment, info_environments, ...)
}
\arguments{
\item{environment}{\code{character} Name of the environment for which climate
data should be extracted.}

\item{info_environments}{\code{data.frame} object with at least the 4 first
columns. \cr
\enumerate{
\item year: \code{numeric} Year label of the environment
\item location: \code{character} Name of the location
\item longitude: \code{numeric} longitude of the environment
\item latitude: \code{numeric} latitude of the environment
\item planting.date: (optional) \code{Date} YYYY-MM-DD
\item harvest.date: (optional) \code{Date} YYYY-MM-DD
\item elevation: (optional) \code{numeric}
\item IDenv: \code{character} ID of the environment (location x year)\cr
}
\strong{The data.frame should contain as many rows as Year x Location
combinations. Example: if only one location evaluated across four years, 4
rows should be present.}}
}
\value{
a data.frame \code{data.frame} with the following columns extracted
from SoilGrids
\enumerate{
\item IDenv \code{character}
\item a list of soil features

}
}
\description{
Function downloading soil data from SoilGrids database
}
\references{
\insertRef{poggio2021soilgrids}{learnMET}
}
\author{
Cathy C. Westhues \email{cathy.jubin@hotmail.com}
}
