% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_solar_radiation.R
\name{get_solar_radiation}
\alias{get_solar_radiation}
\title{Obtain daily solar radiation for an IDenv with package nasapower
derived by NASA from satellite & atmospheric observations.}
\usage{
get_solar_radiation(environment, info_environments, ...)
}
\arguments{
\item{environment}{\code{character} Name of the environment for which climate
data should be extracted.}

\item{info_environments}{\code{data.frame} object with at least the 4 first
columns. \cr
\enumerate{
\item year: \code{numeric} Year label of the environment
\item location: \code{character} Name of the location
\item longitude: \code{numeric} longitude of the environment
\item latitude: \code{numeric} latitude of the environment
\item planting.date: (optional) \code{Date} YYYY-MM-DD
\item harvest.date: (optional) \code{Date} YYYY-MM-DD
\item IDenv: \code{character} ID of the environment (location x year)\cr
}
\strong{The data.frame should contain as many rows as Year x Location
combinations. Example: if only one location evaluated across four years, 4
rows should be present.}}
}
\value{
a data.frame \code{data.frame} with the following columns extracted
from POWER data, according to requested parameters:
\enumerate{
\item longitude \code{numeric}
\item latitude \code{numeric}
\item YEAR \code{numeric}
\item MM \code{integer}
\item DD \code{integer}
\item DOY \code{integer}
\item YYYYMMDD \code{Date}
\item ALLSKY_SFC_SW_DWN \code{numeric}
\item IDenv \code{character} ID environment for which weather data were
downloaded.
\item length.gs \code{difftime} length in days of the growing season
for the environment.
}
}
\description{
Function downloading daily weather data via the package \strong{nasapower}
based on longitude, latitude, planting and harvest date characterizing this
environment.
}
\references{
\insertRef{sparks2018nasapower}{learnMET}
}
\author{
Cathy C. Westhues \email{cathy.jubin@hotmail.com}
}
