% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_splits_processed_with_method.R
\name{get_splits_processed_with_method}
\alias{get_splits_processed_with_method}
\title{Attribute}
\usage{
get_splits_processed_with_method(
  splits,
  method_processing,
  trait,
  geno_data,
  env_predictors,
  info_environments,
  unique_EC_by_geno,
  geno_information,
  use_selected_markers,
  SNPs,
  list_env_predictors,
  include_env_predictors,
  lat_lon_included,
  year_included,
  ...
)
}
\arguments{
\item{splits}{zw4zw}

\item{method_processing}{rze}
}
\value{
jj
}
\description{
Implement trait prediction based on SNP and environmental data
with selection of prediction methods among Machine Learning approaches.

This function should be used to assess the predictive ability according to
a cross-validation scheme determined by the user.
}
\author{
Cathy C. Jubin \email{cathy.jubin@uni-goettingen.de}
}
