% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\docType{data}
\name{info_environments_G2F}
\alias{info_environments_G2F}
\title{Maize experimental multi-environment data sets (Genomes to Fields Initiative)}
\format{
A \code{data.frame} object with the 4 following columns. \cr
\enumerate{
\item year: \code{numeric} Year label of the environment
\item location: \code{character} Name of the location
\item longitude: \code{numeric} longitude of the environment
\item latitude: \code{numeric} latitude of the environment
\item planting.date: \code{Date} Planting date, format "YYYY-MM-DD"
\item harvest.date: \code{Date} Harvest date, format "YYYY-MM-DD"
}
}
\source{
Subset of the publicly available datasets from the Genomes to Fields Initiative. \cr
All data are publicly available at: https://www.genomes2fields.org/resources/,
}
\description{
This data are used as toy data for several functions of
\pkg{learnMET} package. \cr
Field information about 22 environments from the Genomes to Fields project.
}
\references{
\insertRef{mcfarland2020maize}{learnMET}
\insertRef{alkhalifah2018maize}{learnMET}
}
\author{
Cathy C. Westhues \email{cathy.jubin@uni-goettingen.de}
}
\keyword{data}
