% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\docType{data}
\name{info_environments_japonica}
\alias{info_environments_japonica}
\title{Multi-year trial data of rice}
\format{
A \code{data.frame} object with the 4 following columns. \cr
\enumerate{
\item year: \code{numeric} Year label of the environment
\item location: \code{character} Name of the location
\item longitude: \code{numeric} longitude of the environment
\item latitude: \code{numeric} latitude of the environment
}
}
\source{
Data from the INIA's Rice Breeding Program (Uruguay). Data downloaded from publication Monteverde et al. (2019).
}
\description{
This data are used as toy data for several functions of
\pkg{learnMET} package.\cr
Information about environments where \emph{japonica} rice population was grown.
}
\references{
\insertRef{monteverde2019integrating}{learnMET} \
Instituto Nacional de Investigación Agropecuaria (INIA). Programa Arroz. Estación Experimental INIA Treinta y Tres, Uruguay
}
\author{
Cathy C. Westhues \email{cathy.jubin@uni-goettingen.de}
}
\keyword{data}
