% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\docType{data}
\name{intervals_growth_manual_G2F}
\alias{intervals_growth_manual_G2F}
\title{Maize experimental multi-environment data sets (Genomes to Fields Initiative)}
\format{
\code{data.frame} providing the dates at which the field crop enters
a new growth stage, based on estimations \strong{in each environment}.
"P" refers to planting date and should be provided in the third column.
\code{data.frame} with:
\enumerate{
\item year: \code{numeric}
\item location: \code{character}
\item columns 3 and +: \code{numeric} Date (in Days after Planting) at
which the crop enters a new growth stage. "P" refers to the planting date
and should contain 0 as value, "VE" to emergence, etc...
}
\strong{This data.frame is provided as AN EXAMPLE. It does not reflect at
all the true growth stages observed in each individual environment, but the
objective is to provide an example on how information about growth stages
can be given by the user and used to estimate phenological intervals.}
}
\description{
This data are used as toy data for several functions of
\pkg{learnMET} package. \cr
}
\author{
Cathy C. Westhues \email{cathy.jubin@uni-goettingen.de}
}
\keyword{data}
