% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\docType{data}
\name{map_indica}
\alias{map_indica}
\title{Multi-year trial data of rice}
\format{
map \code{data.frame} object with 3 columns.
\enumerate{
\item marker_name \code{character} with marker names
\item chr \code{numeric} with chromosome number
\item pos \code{numeric} with marker position.
}
}
\source{
Data from the INIA's Rice Breeding Program (Uruguay). Data downloaded from publication Monteverde et al. (2019).
}
\description{
This data are used as toy data for several functions of
\pkg{learnMET} package.\cr
Physical map (GBS dataset) for the \emph{indica} population (92,430 markers)
}
\references{
\insertRef{monteverde2019integrating}{learnMET} \
Instituto Nacional de Investigación Agropecuaria (INIA). Programa Arroz. Estación Experimental INIA Treinta y Tres, Uruguay
}
\author{
Cathy C. Westhues \email{cathy.jubin@uni-goettingen.de}
}
\keyword{data}
