% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_effect_per_env_BLINK.R
\name{marker_effect_per_env_BLINK}
\alias{marker_effect_per_env_BLINK}
\title{Compute marker effects per environment with BLINK}
\usage{
marker_effect_per_env_BLINK(
  environment,
  geno,
  pheno,
  map,
  trait,
  nb_folds_cv = 10,
  reps = 2
)
}
\arguments{
\item{environment}{\code{character} indicating the name of the environment
for which marker effects should be computed}

\item{geno}{\code{data.frame} with markers in columns and inviduals in rows.
Typical input is METData$geno.}

\item{pheno}{\code{data.frame} with:
First column: ID genotypes
Second column: year
Third column: location
Subsequent columns: phenotypic traits with names indicated in colnames()
Last column: IDenv (combination LocationxYear)
Typical input is METData$pheno}

\item{trait}{\code{character} Name of the trait under study for which marker
effects should be estimated}
}
\value{
First column contains the marker names.
Second column contains the marker effects in this environment
calculated by cross-validation.
Third column contains the environment name (combination LocationxYear).
}
\description{
Compute marker effects per environment with BLINK
}
