% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_effect_per_env_EN.R
\name{marker_effect_per_env_EN}
\alias{marker_effect_per_env_EN}
\title{Compute marker effects per environment with Elastic Net}
\usage{
marker_effect_per_env_EN(
  geno,
  pheno,
  environment,
  pheno_trait,
  nb_folds_cv = 4,
  reps = 2,
  ...
)
}
\arguments{
\item{geno}{\code{data.frame} with markers in columns and inviduals in rows.
Typical input is \code{METData$geno} from the \code{\link[=create_METData]{create_METData()}} function.}

\item{pheno}{\code{data.frame} with:
First column: ID genotypes
Second column: year
Third column: location
Subsequent columns: phenotypic traits with names indicated in colnames()
Last column: IDenv (combination LocationxYear)
Typical input is \code{METData$pheno} from the \code{\link[=create_METData]{create_METData()}} function.}

\item{environment}{\code{character} indicating the name of the environment
for which marker effects should be computed}

\item{pheno_trait}{\code{character} Name of the trait under study for which
marker effects should be estimated}

\item{nb_folds_cv}{\code{numeric} Number of folds in the CV process}

\item{reps}{\code{numeric} Number of repeats of the k-folds CV}
}
\value{
all_coef_avg \code{data.frame}
\enumerate{
\item First column \code{character} contains the marker names.
\item Second column \code{numeric} the marker effects in this environment
calculated by cross-validation with Elastic Net.
\item Third column \code{character} contains the environment name
(combination LocationxYear).
}
}
\description{
Compute marker effects per environment with Elastic Net
}
\author{
Cathy C. Jubin \email{cathy.jubin@uni-goettingen.de}
}
