% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_effect_per_env_FarmCPU.R
\name{marker_effect_per_env_FarmCPU}
\alias{marker_effect_per_env_FarmCPU}
\title{Compute marker P-values for each environment with FarmCPU (GWAS method).
Multiple testing correction: Benjamini–Hochberg procedure with alpha=0.05.}
\usage{
marker_effect_per_env_FarmCPU(
  geno,
  pheno,
  map,
  environment,
  pheno_trait,
  nb_pcs = 5,
  ...
)
}
\arguments{
\item{geno}{\code{data.frame} with markers in columns and inviduals in rows.

Typical input is METData$geno, after having created a METData object using
create_METData() function}

\item{pheno}{\code{data.frame} with:
First column: ID genotypes
Second column: year
Third column: location
Subsequent columns: phenotypic traits with names indicated in colnames()
Last column: IDenv (combination LocationxYear)

Typical input is METData$pheno, after having created a METData object using
create_METData() function}

\item{map}{\code{data.frame} with 3 columns.
First column \code{character} with marker names
Second column \code{numeric} with chromosome number
Third column \code{numeric} with marker position.

Typical input is METData$map, after having created a METData object using
create_METData() function}

\item{environment}{\code{character} indicating the name of the environment
for which marker effects should be computed}

\item{pheno_trait}{\code{character} Name of the trait under study for which marker
effects should be estimated}

\item{nb_pcs}{\code{numeric} Number of principal components to use in FarmCPU
analyses.}
}
\value{
a \code{list} which contains the following elements:

\describe{
\item{GWAS_results}{\code{data.frame} FarmCPU results for all SNPs}
\item{threshold}{\code{numeric} Cutoff derived from Benjamini-Hochberg
procedure}
\item{selected_markers}{\code{character} Vector containing the names of the
SNPs passing the threshold}
}
}
\description{
Compute marker P-values for each environment with FarmCPU (GWAS method).
Multiple testing correction: Benjamini–Hochberg procedure with alpha=0.05.
}
\author{
Cathy C. Jubin \email{cathy.jubin@uni-goettingen.de}
}
