% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal_fitting_train_test_split.R
\name{ordinal_fitting_train_test_split}
\alias{ordinal_fitting_train_test_split}
\title{Fitting a model on the training set and predicting the test set.}
\usage{
ordinal_fitting_train_test_split(
  split,
  prediction_method = c("xgboost"),
  treatment_ordinal_data = c("ordinal_prediction"),
  seed,
  nb_classes,
  inner_cv_reps = 2,
  inner_cv_folds = 5,
  return_finalized_train_test_sets = F,
  ...
)
}
\arguments{
\item{prediction_method}{\code{character} Prediction method other than
kernel-based methods to use on the predictors defined by the recipe item of
the split object.Options are 'xgboost' for XGBoost, 'rf' for Random Forest.
Default is \code{xgboost}.}

\item{treatment_ordinal_data}{'naive_multiclass','ordinal_prediction',
'regression'}

\item{seed}{\code{integer} Seed value.}

\item{inner_cv_reps}{\code{integer} Number of times to repeat the k-fold
partitioning used for the inner cross-validation for estimation of the best
hyperparameters. Default is 2.}

\item{inner_cv_folds}{\code{integer} Number of partitions of the training set
for the inner cross-validation for estimation of the best hyperparameters.
Default is 5.}

\item{return_finalized_train_test_sets}{a \code{logical} whether the trained
dataset and the test dataset resulting from preprocessing operations
should be returned in the final \code{res_fitted_split} object. Default is
\code{FALSE}.}
}
\value{
a \code{list} object of class \code{res_fitted_split} with the
following items:
\itemize{
\item \strong{predictions_df}: \code{data.frame} with original test dataset with
extra column containing predicted values.
\item \strong{cor_pred_obs}: \code{numeric} Pearson's correlation between predicted
and observed values of the test set.
\item \strong{rmse_pred_obs}: \code{numeric} root mean square error between
predicted and observed values of the test set.
\item \strong{best_hyperparameters}: a \code{tbl_df} giving the tuning parameter
combination with the best performance values which was used to fit the
final model on the training set.
}
}
\description{
Function creating a workflow based on the prediction model chosen (all
methods other than multi-kernel learning), with hyperparameter optimization
using a Bayesian approach on the training set. Once the best hyperparameters
are identified via resampling, the model is fitted on the complete dataset.
Finally the test set is predicted.
}
\author{
Cathy C. Jubin \email{cathy.jubin@uni-goettingen.de}
}
