% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etp_functions.R
\name{penman_monteith_reference_et0}
\alias{penman_monteith_reference_et0}
\title{Calculates reference ET0 based on the Penman-Monteith model (FAO-56 Method)}
\usage{
penman_monteith_reference_et0(
  doy,
  latitude,
  elevation,
  tmin,
  tmax,
  tmean,
  solar_radiation,
  wind_speed,
  rhmean,
  rhmax,
  rhmin,
  tdew,
  use_rh = TRUE
)
}
\description{
This function calculates the potential evapotranspiration rate from
a reference crop canopy (ET0) in mm/d.

For these calculations the
analysis by FAO is followed as laid down in the FAO publication
\verb{Guidelines for computing crop water requirements - FAO Irrigation and drainage paper 56 <http://www.fao.org/docrep/X0490E/x0490e00.htm#Contents>}_
solar_radiation: daily shortwave radiation
}
\author{
Cathy C. Jubin \email{cathy.jubin@uni-goettingen.de}
}
