% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_based_vip.R
\name{permutation_based_vip}
\alias{permutation_based_vip}
\title{Compute variable importance according to the machine learning algorithm used}
\usage{
permutation_based_vip(model, x, y, permutations, predictors, path_plot)
}
\arguments{
\item{permutations}{= 20}

\item{object}{an object of class \code{res_fitted_split}}

\item{type}{\code{model_specific} or \code{model_agnostic}}

\item{unseen_data}{}
}
\description{
Variable importance calculated with permutation approach
}
\author{
Cathy C. Westhues \email{cathy.jubin@uni-goettingen.de}
}
