% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\docType{data}
\name{pheno_G2F}
\alias{pheno_G2F}
\title{Maize experimental multi-environment data sets (Genomes to Fields Initiative)}
\format{
\code{data.frame} object with 6 columns:
\enumerate{
\item geno_ID \code{character} contains the genotype identifiers.
\item year \code{numeric} contains the year of the observation.
\item location \code{character} contains the name of the location.
\item pltht \code{character} plant height (cm)
\item yld_bu_ac \code{character} grain yield (bushels per acre)
\item earht \code{character} ear height (cm)
}
}
\source{
Subset of the publicly available datasets from the Genomes to Fields Initiative. \cr
All data are publicly available at: https://www.genomes2fields.org/resources/,
}
\description{
This data are used as toy data for several functions of
\pkg{learnMET} package. \cr
Phenotypic data for 22 environments from the Genomes to Fields project.
}
\references{
\insertRef{mcfarland2020maize}{learnMET}
\insertRef{alkhalifah2018maize}{learnMET}
}
\author{
Cathy C. Westhues \email{cathy.jubin@uni-goettingen.de}
}
\keyword{data}
