% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\docType{data}
\name{pheno_japonica}
\alias{pheno_japonica}
\title{Multi-year trial data of rice}
\format{
\code{data.frame} object with 7 columns:
\enumerate{
\item geno_ID \code{character} contains the genotype identifiers.
\item year \code{numeric} contains the year of the observation.
\item location \code{character} contains the name of the location.
\item GY \code{character} Grain Yield
\item PHR \code{character} Percentage of Head Rice Recovery
\item GC \code{character} Percentage of Chalky Grain
\item PH \code{character} Plant Height
}
}
\source{
Data from the INIA's Rice Breeding Program (Uruguay). Data downloaded from publication Monteverde et al. (2019).
}
\description{
This data are used as toy data for several functions of
\pkg{learnMET} package.\cr
Phenotypic data for the \emph{japonica} rice population for four traits.
}
\references{
\insertRef{monteverde2019integrating}{learnMET} \cr
Instituto Nacional de Investigación Agropecuaria (INIA). Programa Arroz. Estación Experimental INIA Treinta y Tres, Uruguay
}
\author{
Cathy C. Westhues \email{cathy.jubin@uni-goettingen.de}
}
\keyword{data}
