% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results_cv.R
\name{plot_results_cv}
\alias{plot_results_cv}
\title{bdyy}
\usage{
plot_results_cv(
  fitting_all_splits,
  trait,
  info_environments,
  cv_type,
  cv0_type,
  filename_plot_PA,
  path_folder,
  nb_folds_cv1,
  repeats_cv1,
  nb_folds_cv2,
  repeats_cv2
)
}
\arguments{
\item{fa}{hh}
}
\value{
f
}
\description{
the prediction accuracy is computed as the correlations between the observed
and predicted values within same environments.
}
\author{
Cathy C. Jubin \email{cathy.jubin@uni-goettingen.de}
}
