% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results_vip_functions.R
\name{plot_results_vip_cv}
\alias{plot_results_vip_cv}
\title{Plot variable importance scores}
\usage{
plot_results_vip_cv(
  fitting_all_splits,
  cv_type,
  cv0_type,
  path_folder,
  nb_folds_cv1,
  repeats_cv1,
  nb_folds_cv2,
  repeats_cv2
)
}
\value{
A variable importance plot is saved in the \code{path_folder}.
}
\description{
Internal function of \code{\link[=predict_trait_MET_cv]{predict_trait_MET_cv()}}.\cr
It plots variable importance scores if the \code{compute_vip} option is set to TRUE in \code{\link[=predict_trait_MET_cv]{predict_trait_MET_cv()}}.\cr
Plots are done at the CV scheme level from the , which means that:
\enumerate{
\item If CV0 is evaluated, the plot shows the 40 most important variables according to the predicted element (i.e. site, year or environment).
\item For CV1 and CV2, variable importance plots are based on a average of the importance of each feature over all training/test splits.
}

Variable importance can be calculated based on model agnostic approaches (permutation-based methods, like for \code{stacking_reg_1} or \code{DL_reg}), or
on model-specific methods (gain metric for GBDT methods \code{xgb_reg}).
}
\author{
Cathy C. Westhues \email{cathy.jubin@uni-goettingen.de}
}
