% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_cv0.R
\name{predict_cv0}
\alias{predict_cv0}
\title{Get train/test splits of the phenotypic MET dataset based on CV0.}
\usage{
predict_cv0(pheno_data, cv0_type)
}
\arguments{
\item{pheno_data}{\code{data.frame} Dataset containing phenotypic outcome
data, as well as the predictor variables.}

\item{cv0_type}{\code{character} either \code{leave-one-environment-out},
\code{leave-one-site-out}, \code{leave-one-year-out} or \code{forward-prediction}.}
}
\value{
a \code{cv_object} object which contains the train/test splits of the
CV scheme. Each element of the object corresponds to a \code{split} object with
two elements:
\describe{
\item{training}{\code{data.frame} Dataset with all observations for the
training set.}
\item{test}{\code{data.frame} Dataset with all observations for the test
set.}
}
}
\description{
Get train/test splits of the phenotypic MET dataset based on a
number of random k-folds partitions determined by the user, according to the
type CV0. Creation of the list of train/test splits based on phenotypic data,
so that all the phenotypes from the same environment/year/site appear in the
same fold, according to the type of the CV0 scheme.
}
\author{
Cathy C. Jubin \email{cathy.jubin@uni-goettingen.de}
}
