% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_trait_MET_cv.R
\name{predict_trait_MET_cv}
\alias{predict_trait_MET_cv}
\title{Cross-validation procedure for phenotypic prediction}
\usage{
predict_trait_MET_cv(
  METData,
  trait,
  method = c("xgboost", "kernel_based"),
  use_selected_markers = T,
  geno_information = c("SNPs", "PCs", "PCs_G"),
  num_pcs = 200,
  lat_lon_included = T,
  year_included = F,
  cv_type = c("cv0", "cv1", "cv2"),
  cv0_type = c("leave-one-environment-out", "leave-one-site-out", "leave-one-year-out",
    "forward-prediction"),
  nb_folds_cv1 = 5,
  repeats_cv1 = 50,
  nb_folds_cv2 = 5,
  repeats_cv2 = 50,
  include_env_predictors = T,
  list_env_predictors = NULL,
  plot_PA = T,
  path_plot_PA = "",
  save_processing = F,
  path_folder = NULL,
  ...
)
}
\arguments{
\item{METData}{\code{list} An object created by the initial function of the
package create_METData().}

\item{trait}{\code{character} Name of the trait to predict.}

\item{method}{\code{character} specifying the predictive model to use.}

\item{use_selected_markers}{A \code{Logical} indicating whether to use a
subset of markers obtained from a previous step
(see \code{\link[=function select_markers]{function select_markers()}}).}

\item{geno_information}{indicating how the genotypic
information should be processed to be used in predictions. Options are
\code{SNPs}, \code{PCs} or \code{PCs_G}.}

\item{num_pcs}{\code{}. Default is 200.}

\item{lat_lon_included}{\code{logical} indicates if longitude and latitude
data should be used as numeric predictors. Default is \code{TRUE}}

\item{year_included}{\code{logical} indicates if year factor should be used
as predictor variable. Default is \code{FALSE}.}

\item{cv_type}{= c('cv0', 'cv1', 'cv2')}

\item{cv0_type}{= c( 'leave-one-environment-out','leave-one-site-out',
'leave-one-year-out','forward-prediction'),}

\item{nb_folds_cv1}{= 5}

\item{repeats_cv1}{= 50}

\item{nb_folds_cv2}{= 5,}

\item{repeats_cv2}{= 50}

\item{include_env_predictors}{= T}

\item{list_env_predictors}{A \code{character} vector containing the names
of the environmental predictors. By default \code{NULL}: all environmental
predictors
included in the env_data table of the METData object will be used.
= NULL
@param ...
Arguments passed to the plotting functions named in combo.}

\item{plot_PA}{= T}

\item{path_plot_PA}{= ''}
}
\value{
Object of class....
}
\description{
Implement trait prediction based on SNP and environmental data
with a range of prediction methods (Machine Learning)

This function should be used to assess the predictive ability according to
a cross-validation scheme determined by the user.
}
\author{
Cathy C. Jubin \email{cathy.jubin@uni-goettingen.de}
}
