% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.METData.R
\name{print.summary.METData}
\alias{print.summary.METData}
\title{Print the summary of an object of class METData}
\usage{
\method{print}{summary.METData}(x, ...)
}
\arguments{
\item{object}{of class `summary.METData``}
}
\description{
Print different basic information about the MET
dataset are retrieved from the METData object.
}
