% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_train_test_split_kernel.R
\name{processing_train_test_split_kernel}
\alias{processing_train_test_split_kernel}
\title{Processing and selecting predictors to fit the model}
\usage{
processing_train_test_split_kernel(
  split,
  geno_data,
  geno_information,
  use_selected_markers,
  SNPs,
  list_env_predictors,
  include_env_predictors,
  lat_lon_included = F,
  year_included = F,
  ...
)
}
\arguments{
\item{split}{An object of class \code{split}, corresponding to one element of
the total \code{cv_object} generated by one of the functions \code{\link[=predict_cv0]{predict_cv0()}},
\code{\link[=predict_cv1]{predict_cv1()}}, or \code{\link[=predict_cv2]{predict_cv2()}}, and containing the following items:
\itemize{
\item \strong{training}: \code{data.frame} Training dataset
\item \strong{test}: \code{data.frame} Test dataset
}}

\item{geno_data}{\code{data.frame} A \code{geno} element of an object of class
\code{METData}.}

\item{geno_information}{A \code{character} indicating how the genotypic
information should be processed to be used in predictions. Options are
\code{SNPs}, \code{PCs} or \code{PCs_G}.}

\item{use_selected_markers}{A \code{logical} indicating whether to use a
subset of markers obtained from a previous step (see function
\code{\link[=select_markers]{select_markers()}}).}

\item{SNPs}{A \code{data.frame} with the genotype matrix for the selected
markers of the \code{\link[=select_markers]{select_markers()}} step.}

\item{list_env_predictors}{A \code{character} vector containing the names
of the environmental predictors which should be used in predictions.}

\item{include_env_predictors}{A \code{logical} indicating whether
environmental covariates characterizing each environment should be used in
predictions.}

\item{lat_lon_included}{\code{logical} indicates if longitude and latitude
data should be used as numeric predictors. Default is \code{FALSE}.}

\item{year_included}{\code{logical} indicates if year factor should be used
as predictor variable. Default is \code{FALSE}.}
}
\value{
an object of class \code{split_processed} with the following items:
\itemize{
\item \strong{training}: \code{data.frame} Training set.
\item \strong{test}:\code{data.frame} Test set.
\item \strong{rec_G}:\code{recipe} object with the different steps to process
molecular marker dataset from the training set. Same transformations
applied on the test set.
\item \strong{rec_E}:\code{recipe} object with the different steps to process
environmental data from the training set. Same transformations applied on
the test set.
\item \strong{rec_GE}:\code{recipe} object with the different steps to process
the GxE dataset from the training set. Same transformations applied on
the test set.
}
}
\description{
Processing and selecting predictors to fit the model
}
\author{
Cathy C. Jubin \email{cathy.jubin@uni-goettingen.de}
}
