% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_raw_weather_data.R
\name{qc_raw_weather_data}
\alias{qc_raw_weather_data}
\title{Quality control on daily weather data}
\usage{
qc_raw_weather_data(daily_weather_data)
}
\arguments{
\item{daily_weather_data}{a \code{data.frame} which contains the following
mandatory columns:
\enumerate{
\item longitude \code{numeric}
\item latitude \code{numeric}
\item year \code{numeric}
\item location \code{character}
\item YYYYMMDD \code{Date}
\item IDenv \code{character}
\item DOY \code{integer}
}
Available weather data provided by user must be a subset of the following
weather variable names (not required all included). Colnames must be
respected.
\enumerate{
\item T2M \code{numeric} Daily mean temperature (째C)
\item T2M_MIN \code{numeric} Daily minimum temperature (째C)
\item T2M_MAX \code{numeric} Daily maximum temperature (째C)
\item PRECTOT \code{numeric} Daily total precipitation (mm)
\item daily_solar_radiation \code{numeric} daily solar radiation
(MJ/m^2/day)
\item top_atmosphere_insolation \code{numeric} Top-of-atmosphere
Insolation (MJ/m^2/day)
\item T2MDEW \code{numeric} Dew Point (째C)
}}
}
\value{
a processed \code{data.frame} after quality check with the same
columns as before the QC.
\strong{Values which are outside the range of possible
values are assigned to NA. Warning messages are also thrown if some
observations do not pass either the persistency test or the internal
consistency test. Concerned values are not flagged neither assigned to NA
but we recommend the user to have a second look at the daily weather data
provided in this case.}
}
\description{
This function checks range of values for \code{METData} and implements
various test on the daily weather data (persistence tests, internal
consistency tests).
}
\author{
Cathy C. Jubin \email{cathy.jubin@uni-goettingen.de}
}
