% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapor_pressure.R
\name{sat_vap_pressure}
\alias{sat_vap_pressure}
\title{Formulas to compute saturated vapor pressure deficit}
\usage{
sat_vap_pressure(temp)
}
\arguments{
\item{temp}{\code{numeric} air temperature, °C}
}
\value{
saturated vapor pressure
}
\description{
Formula to calculate saturation vapor pressure from temperature
}
\references{
\insertRef{zotarelli2010step}{learnMET}
}
\author{
Cathy C. Westhues \email{cathy.jubin@hotmail.com}
}
