% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\docType{data}
\name{soil_G2F}
\alias{soil_G2F}
\title{Maize experimental multi-environment data sets (Genomes to Fields Initiative)}
\format{
\code{data.frame} object with 3 columns.
\enumerate{
\item year: \code{numeric} Year label of the environment
\item location: \code{character} Name of the location
\item SandProp: Sand composition (\%)
\item ClayProp: Clay composition (\%)
\item SiltProp: Silt composition (\%)
\item OM: Percentage of organic matter (\%)
}
}
\source{
Subset of the publicly available datasets from the Genomes to Fields Initiative. \cr
All data are publicly available at: https://www.genomes2fields.org/resources/,
}
\description{
This data are used as toy data for several functions of
\pkg{learnMET} package. \cr
Soil data for 22 environments of the maize MET dataset.
}
\references{
\insertRef{mcfarland2020maize}{learnMET} \cr
\insertRef{alkhalifah2018maize}{learnMET}
}
\author{
Cathy C. Westhues \email{cathy.jubin@uni-goettingen.de}
}
\keyword{data}
