% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.METData.R
\name{summary.METData}
\alias{summary.METData}
\title{Summary of an object of class METData}
\usage{
\method{summary}{METData}(object, ...)
}
\arguments{
\item{object}{of class \code{METData}}
}
\value{
object of class \code{summary.METData}
}
\description{
Different basic information about the MET
dataset are retrieved from the METData object.
}
