% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb_multiclass_factor.R
\name{xgb_multiclass_factor}
\alias{xgb_multiclass_factor}
\alias{new_xgb_multiclass_factor}
\alias{validate_xgb_multiclass_factor}
\title{blabla}
\usage{
new_xgb_multiclass_factor(
  split,
  trait,
  geno_data,
  env_predictors,
  info_environments,
  unique_EC_by_geno,
  geno_information,
  use_selected_markers,
  SNPs,
  list_env_predictors,
  include_env_predictors,
  lat_lon_included,
  year_included
)

xgb_multiclass_factor(
  split,
  trait,
  geno_data,
  env_predictors,
  info_environments,
  unique_EC_by_geno,
  geno_information,
  use_selected_markers,
  SNPs,
  list_env_predictors,
  include_env_predictors,
  lat_lon_included,
  year_included
)

validate_xgb_multiclass_factor(x)
}
\description{
blabla
}
