% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb_ordinal.R
\name{xgb_ordinal}
\alias{xgb_ordinal}
\alias{new_xgb_ordinal}
\alias{validate_xgb_ordinal}
\title{xgb ordinal}
\usage{
new_xgb_ordinal(
  split = NULL,
  trait = NULL,
  geno_data = NULL,
  env_predictors = NULL,
  info_environments = NULL,
  geno_information = "SNPs",
  use_selected_markers = F,
  SNPs = NULL,
  include_env_predictors = T,
  list_env_predictors = NULL,
  lat_lon_included = F,
  year_included = F,
  ...
)

xgb_ordinal(
  split,
  trait,
  geno_data,
  env_predictors,
  info_environments,
  geno_information,
  use_selected_markers,
  SNPs,
  list_env_predictors,
  include_env_predictors,
  lat_lon_included,
  year_included,
  ...
)

validate_xgb_ordinal(x, ...)
}
\arguments{
\item{split}{blabla}
}
\description{
xgb ordinal
}
