% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb_reg.R
\name{xgb_reg}
\alias{xgb_reg}
\alias{new_xgb_reg}
\alias{validate_xgb_reg}
\title{blabla}
\usage{
new_xgb_reg(
  split,
  trait,
  geno_data,
  env_predictors,
  info_environments,
  unique_EC_by_geno,
  geno_information,
  use_selected_markers,
  SNPs,
  list_env_predictors,
  include_env_predictors,
  lat_lon_included,
  year_included
)

xgb_reg(
  split,
  trait,
  geno_data,
  env_predictors,
  info_environments,
  unique_EC_by_geno,
  geno_information,
  use_selected_markers,
  SNPs,
  list_env_predictors,
  include_env_predictors,
  lat_lon_included,
  year_included
)

validate_xgb_reg(x)
}
\description{
blabla
}
