% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb_reg_1.R
\name{xgb_reg_1}
\alias{xgb_reg_1}
\alias{new_xgb_reg_1}
\alias{validate_xgb_reg_1}
\title{Processing of a split object to get data ready to be used and fitted with
a \code{xgb_reg_1} (gradient boosted tree) regression model.}
\usage{
new_xgb_reg_1(
  split = NULL,
  trait = NULL,
  geno = NULL,
  env_predictors = NULL,
  info_environments = NULL,
  use_selected_markers = F,
  SNPs = NULL,
  include_env_predictors = T,
  list_env_predictors = NULL,
  type_location_info = "location_factor",
  year_included = F,
  ...
)

xgb_reg_1(
  split,
  trait,
  geno,
  env_predictors,
  info_environments,
  use_selected_markers,
  SNPs,
  list_env_predictors,
  include_env_predictors,
  type_location_info,
  year_included,
  ...
)

validate_xgb_reg_1(x, ...)
}
\arguments{
\item{split}{an object of class \code{split}.
A \code{split} object contains a training and test elements.}

\item{trait}{\code{character} Name of the trait to predict. An ordinal trait
should be encoded as \code{integer}.}

\item{geno}{\code{data.frame} It corresponds to a \code{geno} element
within an object of class \code{METData}.}

\item{env_predictors}{\code{data.frame} It corresponds to the \code{env_data}
element within an object of class \code{METData}.}

\item{info_environments}{\code{data.frame} It corresponds to the
\code{info_environments} element within an object of class \code{METData}.}

\item{use_selected_markers}{A \code{Logical} indicating whether to use a
subset of markers  identified via single-environment GWAS or based on the
table of marker effects obtained via Elastic Net as predictor variables,
when main genetic effects are modeled with principal components. \cr
If \code{use_selected_markers} is \code{TRUE}, the \code{SNPs} argument should be
provided.
\strong{For more details, see \code{\link[=select_markers]{select_markers()}}}}

\item{SNPs}{A \code{data.frame} with the genotype matrix (individuals in rows
and selected markers in columns) for SNPs selected via the
\code{\link[=select_markers]{select_markers()}} function.
\strong{Optional argument, can remain as \code{NULL} if no single markers should
be incorporated as predictor variables in analyses based on PCA
decomposition.}}

\item{include_env_predictors}{A \code{logical} indicating whether
environmental covariates characterizing each environment should be used in
predictions.}

\item{list_env_predictors}{A \code{character} vector containing the names
of the environmental predictors which should be used in predictions.
\strong{By default \code{NULL}: all environmental predictors included in the
env_data table of the \code{METData} object will be used.}}

\item{year_included}{\code{logical} indicates if year factor should be used
as predictor variable. Default is \code{FALSE}.}

\item{lat_lon_included}{\code{logical} indicates if longitude and latitude
data should be used as numeric predictors. Default is \code{FALSE}.}
}
\value{
A \code{list} object of class \code{xgb_reg_1} with the following items:
\describe{
\item{training}{\code{data.frame} Training set after partial processing}
\item{test}{\code{data.frame} Test set after partial processing}
\item{rec}{A \code{recipe} object, specifying the remaining processing
steps which are implemented when a model is fitted on the training set
with a recipe.}
}
}
\description{
The function processes a split object (training + test sets), according to
the configuration set by the user. For instance, genomic information is
incorporated according to the option set by the user.  A list of specific
environmental covariables to use can be provided.\cr

A recipe is created using the package \code{recipes}, to specify additional
preprocessing steps, such as standardization based on the training set, with
same transformations used on the test set. Variables with null variance are
removed. If year effect is included, it is converted to dummy variables.\cr
Further fitting on the training set with a gradient boosting model (see
function \code{\link[=fit_cv_split.xgb_reg_1]{fit_cv_split.xgb_reg_1()}})).
}
\references{
\insertRef{wickham2019welcome}{learnMET}
\insertRef{tidymodels}{learnMET}
}
