% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_new_METData.R
\name{add_new_METData}
\alias{add_new_METData}
\alias{new_add_new_METData}
\alias{validate_create_METData}
\title{Create a multi-environment trials data object}
\usage{
new_add_new_METData(
  METData_training,
  pheno_new = NULL,
  geno_new = NULL,
  compute_climatic_ECs = FALSE,
  info_environments_to_predict = NULL,
  env_data_manual = NULL,
  raw_weather_data = NULL,
  ...
)

add_new_METData(
  METData_training,
  pheno_new = NULL,
  geno_new = NULL,
  compute_climatic_ECs = FALSE,
  info_environments_to_predict = NULL,
  env_data_manual = NULL,
  raw_weather_data = NULL,
  ...
)

validate_create_METData(x, ...)
}
\arguments{
\item{METData_training}{Will be used as training set.}

\item{pheno_new}{\code{data.frame} object with at least 3 columns.
\enumerate{
\item geno_ID \code{character} contains the genotype identifiers.
\item year \code{numeric} contains the year of the observation.
\item location \code{character} contains the name of the location.
}
Basically, this \code{data.frame} only indicates which genotypes should be
predicted in which environments. No columns with phenotypic values is
expected.}

\item{info_environments_to_predict}{\code{data.frame} object with at least the 4 first
columns. \cr
\enumerate{
\item year: \code{numeric} Year label of the environment
\item location: \code{character} Name of the location
\item longitude: \code{numeric} longitude of the environment
\item latitude: \code{numeric} latitude of the environment
\item planting.date: (optional) \code{Date} YYYY-MM-DD
\item harvest.date: (optional) \code{Date} YYYY-MM-DD \cr
}
\itemize{
\item \strong{The data.frame should contain as many rows as Year x Location
combinations. Example: if only one location evaluated across four years, 4
rows should be present.}
\item \strong{The fifth and sixth columns (planting.date and harvest.date) are
required only if the user wants to download weather data with the
package (setting argument \code{ECs_computed = T} in \code{\link[=create_METData]{create_METData()}} and using
subsequently the function \code{\link[=get_ECs]{get_ECs()}}).}
}}

\item{scenario_weather_data}{\code{character}. Options are:
\itemize{
\item \strong{\code{manually}}: names of columns in the \code{env_data_manual} data.frame should be the
same as in METData_training$env_data.
\item \strong{\code{use_real_EC}}: environmental variables will be estimated for a set of "past" environments based on
information provided in info_environments_to_predict.
using the \code{\link[=use_real_EC]{use_real_EC()}} function. This assumes that the environmental data can be
retrieved and are from the past. only possible if...
different according to compute_EC.
\item \strong{\code{mean_previous_years}}: ECs computed based on average over 5 last years
of the location indicated in info_environments_to_predict
use of the location data from the location in the training set (implies
that ) if computed_ECs== TRUE in METData_training.
give fictive harvest date --> example
\item \strong{\code{previous_year}}:
}}

\item{location_to_use}{\code{character}. Default is NULL.}

\item{env_data}{\code{data.frame} can be let as NULL by user, if no
environment data provided as input. Otherwise, a \code{data.frame} should
be provided.
\strong{The data.frame should contain as many rows as the phenotypic
dataset \code{data.frame}.} \cr
Columns should be:
\enumerate{
\item geno_ID \code{character} with genotype ID
\item year \code{numeric} with the year label
\item location \code{character} with the location character
}
Columns 4 and + should be numeric and contain the environmental covariates.
\cr
\itemize{
\item \strong{Providing env_data  and setting \code{compute_climatic_ECs = T} is
possible. For instance, the user can have some information regarding the
soil composition (\\% \% clay, sand, silt, organic matter content).
A disease status can also be encoded as categorical variable if it affected
some environments. In addition to these covariates, weather-based
covariates will be computed if \code{compute_climatic_ECs = T} with the package if they
were not available to provide as input by the user.}
}}
}
\value{
A formatted \code{list} of class \code{METData} which contains the
following elements:
\itemize{
\item \strong{geno}: \code{matrix} with genotype values of phenotyped individuals.
\item \strong{map}: \code{data.frame} with genetic map.
\item \strong{pheno}: \code{data.frame} with phenotypic trait values.
\item \strong{compute_EC_by_geno}: \code{logical} indicates if environmental
covariates should be later computed.
\item \strong{env_data}: \code{data.frame} with the environmental covariates per
environment (and if genotype-specific, per genotype).
\item \strong{info_environments}: \code{data.frame} contains basic information on
each environment.
}
}
\description{
This function combines all types of data sources (genotypic, phenotypic,
environmental data, information about the environments...) in a single data
object of class \code{METData}.
}
\examples{

}
\author{
Cathy C. Jubin \email{cathy.jubin@uni-goettingen.de}
}
