% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dalex_ale_plots.R
\name{ale_plot_split}
\alias{ale_plot_split}
\alias{ale_plot_split.default}
\alias{ale_plot_split.fitted_xgb_reg_1}
\title{ALE plots feature-wise with DALEX package
Wrapper functions}
\usage{
ale_plot_split(object, ...)

\method{ale_plot_split}{default}(object, ...)

\method{ale_plot_split}{fitted_xgb_reg_1}(object, variable, path_plot)
}
\arguments{
\item{object}{an object of class \code{res_fitted_split}}

\item{variable}{environmental or genomic variable (PC derived from SNPs
matrix or single SNP) for which the ALE plot should be calculated}
}
\description{
Makes an ALE plot for a set of predictor variable of the model.
}
\references{
\insertRef{dalex}{learnMET}
}
\author{
Cathy C. Westhues \email{cathy.jubin@uni-goettingen.de}
}
