% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_PCA.R
\name{apply_pca}
\alias{apply_pca}
\title{Data dimensionality reduction using PCA on a split object.}
\usage{
apply_pca(split, geno, threshold = 0.95, ...)
}
\arguments{
\item{split}{An object of class \code{split}, corresponding to one element of
the total \code{cv_object} generated by one of the functions \code{\link[=predict_cv0]{predict_cv0()}},
\code{\link[=predict_cv00]{predict_cv00()}}, \code{\link[=predict_cv1]{predict_cv1()}}, or \code{\link[=predict_cv2]{predict_cv2()}},
and containing the following items:
\itemize{
\item \strong{training}: \code{data.frame} Training dataset
\item \strong{test}: \code{data.frame} Test dataset
}}

\item{geno}{\code{data.frame} It corresponds to a \code{geno} element
within an object of class \code{METData}.}

\item{threshold}{\code{numeric} A fraction of the total variance that
should be covered by the components}
}
\value{
pc_values A \code{data.frame} containing the principal components
in columns and the names of all lines used in the study is contained in the
first column 'geno_ID'. PCs for the lines present in the test set were
computed based on the transformation done on the training set.
}
\description{
Fit PCA on the training set and apply the same transformation to the test
set. The goal is to use principal components in prediction models as a
smaller number of variables instead of all the marker predictors.
}
\author{
Cathy C. Westhues \email{cathy.jubin@hotmail.com}
}
