% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_EC.R
\name{compute_EC_fixed_number_windows}
\alias{compute_EC_fixed_number_windows}
\title{Compute ECs based on a fixed number of time windows.}
\usage{
compute_EC_fixed_number_windows(
  table_daily_W = x,
  nb_windows_intervals,
  base_temperature = 10,
  method_GDD_calculation = c("method_b"),
  ...
)
}
\arguments{
\item{table_daily_W}{\code{data.frame} Object returned by the function
get_daily_tables_per_env()}

\item{nb_windows_intervals}{\code{numeric} Number of time windows covering
the growing season length (common number of time windows across all
environments).}

\item{base_temperature}{\code{numeric} Base temperature (crop growth assumed
to be null below this value.). Default is 10.}

\item{method_GDD_calculation}{\code{character} Method used to compute the
GDD value, with one out of \code{method_a} or \code{method_b}. \cr
\code{method_a}: No change of the value of \eqn{T_{min}}.
GDD = \eqn{max (\frac{T_{min}+T_{max}}{2} - T_{base},0)}. \cr
\code{method_b}: If \eqn{T_{min}} < \eqn{T_{base}}, change \eqn{T_{min}}
to \eqn{T_{min}} = \eqn{T_{base}}. \cr
Default \code{method_b}.}
}
\value{
An object of class \code{data.frame} with
10 x number_total_fixed_windows + 1 last column (IDenv):
\enumerate{
\item mean_TMIN: number_total_fixed_windows columns, indicating the
average minimal temperature over the respective time window.
\item mean_TMAX: number_total_fixed_windows columns, indicating the
average maximal temperature over the respective time window.
\item mean_TMEAN: number_total_fixed_windows columns, indicating the
average mean temperature over the respective time window.
\item freq_TMAX_sup30: number_total_fixed_windows columns, indicating the
frequency of days with maximum temperature over 30°C over the respective
time window.
\item freq_TMAX_sup35: number_total_fixed_windows columns, indicating the
frequency of days with maximum temperature over 35°C over the respective
time window.
\item sum_GDD: number_total_fixed_windows columns, indicating the
growing degree days over the respective time window.
\item sum_PTT: number_total_fixed_windows columns, indicating the
accumulated photothermal time over the respective time window.
\item sum_P: number_total_fixed_windows columns, indicating the
accumulated precipitation over the respective time window.
\item freq_P_sup10: number_total_fixed_windows columns, indicating the
frequency of days with total precipitation superior to 10 mm over the
respective time window.
\item sum_solar_radiation: number_total_fixed_windows columns, indicating
the accumulated incoming solar radiation over the respective time window.
\item IDenv \code{character} ID of the environment (Location_Year)
}
@author Cathy C. Jubin \email{cathy.jubin@uni-goettingen.de}
@export
}
\description{
Compute the environmental covariates based on the daily weather
table of an environment (Year x Location), and over a fixed number of time
windows, which is common across all environments. The length of time windows
in days in each environment is determined by dividing the total length of
the growing season of this environment by the number of windows to use.
Each EC is then computed over a fixed number of days, given by
the calculated duration_time_window_days.
}
