% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_METData.R
\name{create_METData}
\alias{create_METData}
\title{Create a multi-environment trials data object}
\usage{
create_METData(
  geno = NULL,
  pheno = NULL,
  info_environments = NULL,
  map = NULL,
  env_data = NULL,
  unique_EC_by_geno = FALSE,
  compute_ECs = FALSE,
  filtering_markers = TRUE
)
}
\arguments{
\item{geno}{\code{numeric} genotype values stored in a \code{matrix} or
\code{data.frame} which contains the geno_ID as row.names and markers as
columns.}

\item{pheno}{\code{data.frame} object with at least 4 columns.
\enumerate{
\item geno_ID \code{character} contains the genotype identifiers.
\item year \code{numeric} contains the year of the observation.
\item location \code{character} contains the name of the location.
}
From the fourth column on: each column \code{numeric} contains phenotypic
values for a phenotypic trait observed in a combination Year x Location.
Names of the traits can be provided as column names.
\cr
\itemize{
\item \strong{The geno_ID must be the same as the row.names in the geno object.
}
\cr
\item \strong{For genotypes to be predicted for real case scenarios (e.g. no
phenotypic data available), fill with NA in the column containing the
phenotypic values of the trait.}
\cr
}}

\item{info_environments}{\code{data.frame} object with at least the 4 first
columns. \cr
\enumerate{
\item year: \code{numeric} Year label of the environment
\item location: \code{character} Name of the location
\item longitude: \code{numeric} longitude of the environment
\item latitude: \code{numeric} latitude of the environment
\item planting.date: (optional) \code{Date} YYYY-MM-DD
\item harvest.date: (optional) \code{Date} YYYY-MM-DD \cr
}
\itemize{
\item \strong{The data.frame should contain as many rows as Year x Location
combinations. Example: if only one location evaluated across four years, 4
rows should be present.}
\item \strong{The fifth and sixth columns (planting.date and harvest.date) are
required only if the user wants to download weather data with the
package (setting argument \code{compute_ECs = T} in \code{\link[=create_METData]{create_METData()}} and using
subsequently the function \code{\link[=get_ECs]{get_ECs()}}).}
}}

\item{map}{\code{data.frame} object with 3 columns.
\enumerate{
\item marker_name \code{character} with marker names
\item chr \code{numeric} with chromosome number
\item pos \code{numeric} with marker position.
}
\strong{Map object not absolutely required}.}

\item{env_data}{\code{data.frame} can be let as NULL by user, if no
environment data provided as input. Otherwise, a \code{data.frame} should
be provided:
Two types of \code{data.frame} can be provided:
\itemize{
\item \strong{Type 1}: if \code{unique_EC_by_geno = FALSE}. Each environmental covariate
characterizes all phenotypes within a complete environment (e.g. ECs are
not computed individually for each variety within an environment).
\strong{The data.frame should contain as many rows as
the info_environments \code{data.frame}.} \cr
Columns should contain:
\enumerate{
\item year \code{numeric} with the year label
\item location \code{character} with the location character
}
Columns 3 and + should be numeric and contain the environmental covariates.
\cr
}

OR \cr
\itemize{
\item \strong{Type 2}: if \code{unique_EC_by_geno = TRUE}. Each environmental covariate
is computed specifically for an environment AND for a genotype (e.g. ECs
are computed individually for each variety within an environment).
\strong{The data.frame should contain as many rows as the phenotypic
dataset \code{data.frame}.} \cr
Columns should contain:
\enumerate{
\item geno_ID \code{character} with genotype ID
\item year \code{numeric} with the year label
\item location \code{character} with the location character
}
Columns 4 and + should be numeric and contain the environmental covariates.
\cr
\item \strong{Providing env_data  and still setting \code{compute_ECs = T} is
possible. For instance, the user can have some information regarding the
soil composition (\\% \% clay, sand, silt, organic matter content).
A disease status can also be encoded as categorical variable if it affected
some environments. In addition to these covariates, weather-based
covariates will be computed if \code{compute_ECs = T} with the package if they
were not available to provide as input by the user.}
}}

\item{unique_EC_by_geno}{\code{logical} indicates if the environmental
covariates contained in env_data are also genotype-specific (dependent on
the phenology, for instance) or unique for a complete environment. Default
is \code{FALSE}.}

\item{compute_ECs}{\code{logical} indicates if environmental covariates
should be computed in further steps using the function \code{\link[=get_ECs]{get_ECs()}}. Default
is \code{FALSE}. \cr
\strong{Set compute_ECs = \code{TRUE} if user wants to use weather data
acquired with this package.}}

\item{filtering_markers}{\code{logical} indicator if a single-environment
QTL detection step (performed with GWAS or ElasticNet using the function
\code{\link[=select_markers]{select_markers()}}) should be conducted, to identify a subset of markers
representing potential environment-specific QTL effects, for which
interactions with environmental covariates can be further investigated.
Default is \code{TRUE}.}
}
\value{
A formatted \code{list} of class \code{METData} which contains the
following elements:
\itemize{
\item \strong{geno}: \code{matrix} with genotype values of phenotyped individuals.
\item \strong{map}: \code{data.frame} with genetic map.
\item \strong{pheno}: \code{data.frame} with phenotypic trait values.
\item \strong{compute_EC_by_geno}: \code{logical} indicates if environmental
covariates should be later computed.
\item \strong{env_data}: \code{data.frame} with the environmental covariates per
environment (and if genotype-specific, per genotype).
\item \strong{info_environments}: \code{data.frame} contains basic information on
each environment.
\item \strong{unique_EC_by_geno}: \code{logical} to indicate if the EC is genotype-
specific.
\item \strong{filtering_markers}: \code{logical} indicates if a filtering marker
step should be applied in further steps
}
}
\description{
This function combines all types of data sources (genotypic, phenotypic,
environmental data, information about the environments...) in a single data
object of class \code{METData}.
}
\examples{

data(geno_G2F)
data(pheno_G2F)
data(map_G2F)
data(info_environments_G2F)
METdata_G2F <- create_METData(geno=geno_G2F,pheno=pheno_G2F,map=map_G2F,env_data = NULL,compute_ECs = F,info_environments = info_environments_G2F)

data(geno_indica)
data(map_indica)
data(pheno_indica)
data(info_environments_indica)
data(env_data_indica)
METdata_indica <- create_METData(geno=geno_indica,pheno=pheno_indica,env_data = env_data_indica,unique_EC_by_geno = F,compute_ECs = F,info_environments = info_environments_indica,map = map_indica)

data(geno_japonica)
data(map_japonica)
data(pheno_japonica)
data(info_environments_japonica)
data(env_data_japonica)
METdata_japonica <- create_METData(geno=geno_japonica,pheno=pheno_japonica,env_data = env_data_japonica,unique_EC_by_geno = F,compute_ECs = F,info_environments = info_environments_japonica,map = map_japonica)

}
\author{
Cathy C. Jubin \email{cathy.jubin@uni-goettingen.de}
}
