% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_METData.R
\name{create_METdata}
\alias{create_METdata}
\title{Create a multi-environment trials data object}
\usage{
create_METdata(
  geno = NULL,
  pheno = NULL,
  map = NULL,
  env_data = NULL,
  unique_EC_by_geno = FALSE,
  compute_ECs = FALSE,
  info_environments = NULL,
  filtering_markers = TRUE
)
}
\arguments{
\item{geno}{\code{numeric} genotype values stored in a \code{matrix} or
\code{data.frame} which contains the geno_ID as row.names and markers as
columns.}

\item{pheno}{\code{data.frame} with at least 4 columns.
First column "geno_ID"  \code{character} contains the genotype identifiers.
\strong{The geno_ID must be the same to the row.names in the geno object}
\strong{For genotypes to be predicted (only geno data, no pheno values),
fill with NA in pheno}
Second column "year"  \code{numeric} contains the year for the observation
Third column "location" \code{character} contains the name of the location
Fourth column and + \code{numeric} contain the phenotypic values for
different traits. Names of the traits can be provided as col.names.}

\item{map}{\code{data.frame} with 3 columns.
First column \code{character} with marker names
Second column \code{numeric} with chromosome number
Third column \code{numeric} with marker position.
\strong{not absolutely required}.}

\item{env_data}{\code{data.frame} can be let as NULL by user, if no
environment data already present.
If env_data provided, and if unique_EC_by_geno, the data.frame should
contain as many rows as the pheno data.frame.
First column \code{character} with genotype identifiers
Second column \code{numeric} with the year label
Third column \code{character} with the location character
Columns 4+ should be numeric and contain the environmental covariates
provided by the user.
If unique_EC_by_geno==FALSE, the data.frame should contain as many rows as
the info_environments data.frame:
First column \code{numeric} with the year label
Second column \code{character} with the location character
Columns 3+ should be numeric and contain the environmental covariates}

\item{unique_EC_by_geno}{\code{Logical} indicates if the environmental
covariates contained in env_data are also genotype-specific (dependent on
the phenology, for instance) or unique for a whole environment.}

\item{compute_ECs}{\code{Logical} indicates if environmental covariates
should be computed in further steps.}

\item{info_environments}{\code{data.frame} with at least 4 columns.
First column \code{numeric} with the year label
Second column \code{character} with the location
Third column \code{numeric} with the longitude
Fourth column \code{numeric} with the latitude

Additional (optional) columns, required if the user wants to download
weather data with the package (via argument compute_ECs = T):
Fifth column \code{Date} planting.date as YYYY-MM-DD
Sixth column \code{Date} harvest.date as YYYY-MM-DD
\strong{The data.frame should contain as many rows as Year x Location
combinations. Example: if only one location used in the analyses for four years,
4 rows should be present (same information with only the value in column year
changing).}}

\item{filtering_markers}{\code{Logical} indicator if the number of markers
to use in the machine-learning based predictions should be pre-handled
in further steps (elasticnet, PCA...)}
}
\value{
a \code{list} of class \code{METData} which contains the following elements
\itemize{

\item{geno}{\code{matrix} with genotype values of phenotyped individuals.}

\item{map}{\code{data.frame} with genetic map.}

\item{pheno}{\code{data.frame} with phenotypic trait values.}

\item{compute_EC_by_geno}{\code{Logical} indicates if environmental
covariates should be later computed.}

\item{env_data}{\code{data.frame} with the environmental covariates per
environment (and if genotype-specific, per genotype).}

\item{info_environments}{\code{data.frame} contains basic information on
each environment.}

\item{unique_EC_by_geno}{\code{Logical} to indicate if the EC is genotype-
specific.}

\item{filtering_markers}{\code{Logical} indicates if a filtering marker step
should be applied in further steps}
}
}
\description{
Create a multi-environment trials data object
}
\examples{

data(geno_G2F)
data(pheno_G2F)
data(info_environments_G2F)
data(env_data_G2F)
METdata_G2F <- create_METdata(geno=geno_G2F,pheno=pheno_G2F,env_data = env_data_G2F,unique_EC_by_geno = T,compute_ECs = F,info_environments = info_environments_G2F)

data(geno_indica)
data(map_indica)
data(pheno_indica)
data(info_environments_indica)
data(env_data_indica)
METdata_indica <- create_METdata(geno=geno_indica,pheno=pheno_indica,env_data = env_data_indica,unique_EC_by_geno = F,compute_ECs = F,info_environments = info_environments_indica,map = map_indica)

data(geno_japonica)
data(map_japonica)
data(pheno_japonica)
data(info_environments_japonica)
data(env_data_japonica)
METdata_japonica <- create_METdata(geno=geno_japonica,pheno=pheno_japonica,env_data = env_data_japonica,unique_EC_by_geno = F,compute_ECs = F,info_environments = info_environments_japonica,map = map_japonica)

}
