% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cv_split.R
\name{fit_cv_split}
\alias{fit_cv_split}
\alias{fit_cv_split.default}
\alias{fit_cv_split.xgb_reg}
\alias{fit_cv_split.xgb_ordinal}
\alias{fit_cv_split.DL_reg}
\alias{fit_cv_split.svm_stacking_reg}
\title{fit_cv_split}
\usage{
fit_cv_split(object, ...)

\method{fit_cv_split}{default}(x, ...)

\method{fit_cv_split}{xgb_reg}(object, seed, inner_cv_reps = 1, inner_cv_folds = 3, ...)

\method{fit_cv_split}{xgb_ordinal}(object, seed, inner_cv_reps = 2, inner_cv_folds = 5, ...)

\method{fit_cv_split}{DL_reg}(object, seed, inner_cv_reps = 1, inner_cv_folds = 3, ...)

\method{fit_cv_split}{svm_stacking_reg}(
  object,
  seed,
  inner_cv_reps = 2,
  inner_cv_folds = 5,
  kernel_G = "rbf",
  kernel_E = "rbf",
  kernel_GE = "rbf",
  ...
)
}
\arguments{
\item{object}{It must be an object of class}

\item{fit}{CV object}
}
\description{
for all CV schemes, the prediction accuracy is computed as the correlations between the observed
and predicted values within same environments.
}
\author{
Cathy C. Jubin \email{cathy.jubin@uni-goettingen.de}
}
