% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_DL_reg.R, R/fit_cv_split.R,
%   R/fit_svm_stacking_reg.R, R/fit_xgb_ordinal.R, R/fit_xgb_reg.R
\name{fit_cv_split.DL_reg}
\alias{fit_cv_split.DL_reg}
\alias{fit_cv_split}
\alias{fit_cv_split.default}
\alias{fit_cv_split.svm_stacking_reg}
\alias{fit_cv_split.xgb_ordinal}
\alias{fit_cv_split.xgb_reg}
\title{S3 method used to fit an object of class \code{xgb_reg}, \code{xgb_ordinal},
\code{svm_stacking_reg} or \code{DL_reg}.}
\usage{
\method{fit_cv_split}{DL_reg}(object, seed, inner_cv_reps = 1, inner_cv_folds = 3, ...)

fit_cv_split(object, ...)

\method{fit_cv_split}{default}(x, ...)

\method{fit_cv_split}{svm_stacking_reg}(
  object,
  seed,
  inner_cv_reps = 2,
  inner_cv_folds = 5,
  kernel_G = "rbf",
  kernel_E = "rbf",
  kernel_GE = "rbf",
  ...
)

\method{fit_cv_split}{xgb_ordinal}(object, seed, inner_cv_reps = 2, inner_cv_folds = 5, ...)

\method{fit_cv_split}{xgb_reg}(object, seed, inner_cv_reps = 1, inner_cv_folds = 3, ...)
}
\arguments{
\item{object}{an object of class \code{xgb_reg}, \code{xgb_ordinal},
\code{svm_stacking_reg} or \code{DL_reg}.}
}
\description{
S2 dispatching method for objects of class \code{xgb_reg}, \code{xgb_ordinal},
\code{svm_stacking_reg} or \code{DL_reg}.
}
\author{
Cathy C. Jubin \email{cathy.jubin@uni-goettingen.de}
}
