% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\docType{data}
\name{geno_indica}
\alias{geno_indica}
\title{Multi-year trial data of rice}
\format{
\code{numeric} matrix with genotype values stored in a \code{matrix} or
\code{data.frame} for the \emph{indica} rice population. Names of lines (geno_ID) are given as row.names and markers as
columns.
Dimensions are \code{n}=327 and \code{p}=92,430.
}
\source{
Data from the INIA's Rice Breeding Program (Uruguay). Data downloaded from publication Monteverde et al. (2019).
}
\description{
This data are used as toy data for several functions of
\pkg{learnMET} package.
}
\references{
\insertRef{monteverde2019integrating}{learnMET} \cr
Instituto Nacional de Investigación Agropecuaria (INIA). Programa Arroz. Estación Experimental INIA Treinta y Tres, Uruguay
}
\author{
Cathy C. Westhues \email{cathy.jubin@uni-goettingen.de}
}
\keyword{data}
