% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ECs.R
\name{get_ECs}
\alias{get_ECs}
\title{Compute environmental covariates for each environment of the MET dataset.}
\usage{
get_ECs(
  info_environments,
  raw_weather_data = NULL,
  method_ECs_intervals = "fixed_nb_windows_across_env",
  length_minimum_gs = NULL,
  save_daily_weather_tables = F,
  path_data = NULL,
  crop_model = NULL,
  nb_windows_intervals = 10,
  duration_time_window_days = 10,
  base_temperature = 10,
  clustering_climate_data = T,
  ...
)
}
\arguments{
\item{info_environments}{a \code{data.frame} with the following columns}

\item{method_ECs_intervals}{GDD}

\item{save_daily_weather_tables}{\code{logical} indicates whether the
daily weather tables should be saved. Default is \code{TRUE}.}

\item{path_data}{\code{character} Path of the folder where a
RDS object will be created to save the daily weather tables if saved. (Do
not use a Slash after the name of the last folder.)}

\item{nb_windows_intervals}{\code{numeric} Number of time windows covering
the growing season length (common number of time windows across all
environments). Default is 5.}

\item{...}{Arguments passed to the \code{\link[=compute_EC]{compute_EC()}} function.
\strong{Not all of the aforementioned weather variables need to be
provided. Weather variables which are not provided and needed to compute
environmental covariables will be retrieved from NASA POWER and merged to
the weather data given by the user. If these environmental covariables
should not be used in predictions, the user can specify in the prediction
function the list of environmental covariables to use with the argument
list_env_predictors.}
\strong{data are imputed if missing or assigned to NA after QC}}

\item{fixed_length_time_windows_across_env}{\code{logical} indicates if the
growing season lengths should be divided in non-overlapping time windows of
fixed lengths (in days) across all environments.
This implies that the total number of time windows, which need to be common
across all environments, is determined by the shortest growing season
included in the MET environments. This further implies that the total
growing season may not be covered by the environmental predictors for
the longest growing seasons.\cr
Default is \code{TRUE}.}

\item{fixed_nb_windows_across_env}{\code{logical} indicates if the
growing season lengths should be divided in a fixed number of
non-overlapping windows, which fully cover the growing season of each
environment. This means that these time windows might not be of same length
across environments (but always of same length in one environment). \cr
Default is \code{FALSE}.}

\item{customized_growth_intervals}{\code{logical} indicates if the growth
intervals to use are given by user and should be used to define time
windows. \cr
Default is \code{FALSE}.}
}
\value{
A \code{data.frame} object containing the weather-based environmental
covariates.
}
\description{
This function enables to retrieve daily weather data from NASA POWER source
for each environment and derive environmental covariates over non-overlapping
time windows, which can be defined in various ways by the user.
The user can also provide own daily weather data, even for only part of the
total number of environments. For the remaining environments, weather data
will be retrieved using the NASA POWER query.
}
\references{
\insertRef{sparks2018nasapower}{learnMET}
}
\author{
Cathy C. Westhues \email{cathy.jubin@uni-goettingen.de}
}
