% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_splits_processed_with_method.R
\name{get_splits_processed_with_method}
\alias{get_splits_processed_with_method}
\title{Attribute a processing method for each list of training/test splits}
\usage{
get_splits_processed_with_method(
  splits,
  prediction_method,
  trait,
  geno,
  env_predictors,
  info_environments,
  use_selected_markers,
  SNPs,
  list_env_predictors,
  include_env_predictors,
  lat_lon_included,
  year_included,
  ...
)
}
\arguments{
\item{splits}{an object of class \code{cv_object}}

\item{prediction_method}{}
}
\value{
a list of objects of the class of the chosen prediction method
}
\description{
Attribute the function for processing the dataset according to the processing parameters.
}
\author{
Cathy C. Westhues \email{cathy.jubin@uni-goettingen.de}
}
