% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_effect_per_env_FarmCPU.R
\name{marker_effect_per_env_FarmCPU}
\alias{marker_effect_per_env_FarmCPU}
\title{Compute marker P-values for each environment with GWAS.}
\usage{
marker_effect_per_env_FarmCPU(
  geno,
  pheno,
  map,
  environment,
  pheno_trait,
  nb_pcs = 5,
  ...
)
}
\arguments{
\item{geno}{\code{data.frame}with markers in columns and inviduals in rows.
Typical input is \code{METData$geno} from the \code{\link[=create_METData]{create_METData()}} function.}

\item{pheno}{\code{data.frame} object with at least 4 columns.
\enumerate{
\item geno_ID \code{character} contains the genotype identifiers.
\item year \code{numeric} contains the year of the observation.
\item location \code{character} contains the name of the location.
}
\itemize{
\item \strong{From the fourth column on: each column \code{numeric} contains
phenotypic values for a phenotypic trait observed in a combination
Year x Location. Names of the traits can be provided as column names.}
\item \strong{Last column: IDenv (combination LocationxYear)}
}

Typical input is \code{METData$pheno} from the \code{\link[=create_METData]{create_METData()}} function.}

\item{map}{\code{data.frame} with 3 columns.
\enumerate{
\item marker \code{character} with marker names
\item chr \code{numeric} with chromosome number
\item pos \code{numeric} with marker position.
}
Typical input is METData$map from the \code{\link[=create_METData]{create_METData()}} function.}

\item{environment}{\code{character} indicating the name of the environment
for which marker effects should be computed}

\item{pheno_trait}{\code{character} Name of the trait under study for which marker
effects should be estimated}

\item{nb_pcs}{\code{numeric} Number of principal components to use in FarmCPU
analyses.}
}
\value{
a \code{list} which contains the following elements:
\describe{
\item{GWAS_results}{\code{data.frame} FarmCPU results for all SNPs}
\item{threshold}{\code{numeric} Cutoff derived from Benjamini-Hochberg
procedure}
\item{selected_markers}{\code{character} Vector containing the names of
the SNPs passing the threshold}
}
}
\description{
GWAS method implemented: FarmCPU from GAPIT3.
Multiple testing correction: Benjamini–Hochberg procedure with alpha=0.05.
}
\details{
For more information about the GWAS method, please consult the original
publication from the authors:
Liu, Xiaolei, et al. "Iterative usage of fixed and random effect models for
powerful and efficient genome-wide association studies."
PLoS genetics 12.2 (2016): e1005767.
\url{https://doi.org/10.1371/journal.pgen.1005767}
}
\author{
Cathy C. Westhues \email{cathy.jubin@hotmail.com}
}
