% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_cv2.R
\name{predict_cv2}
\alias{predict_cv2}
\title{Get train/test splits of the phenotypic MET dataset based on CV2.}
\usage{
predict_cv2(pheno_data, nb_folds, reps, seed)
}
\arguments{
\item{pheno_data}{\code{data.frame} Dataset containing phenotypic outcome
data, as well as the predictor variables}

\item{nb_folds}{\code{numeric} Number of folds in the CV process}

\item{reps}{\code{numeric} Number of repeats of the k-folds CV}
}
\value{
a \code{cv_object} object which contains nb_folds x reps elements.
Each element of the object corresponds to a \code{split} object with two
elements:
\describe{
\item{training}{\code{data.frame} Dataset with all observations for the
training set.}
\item{test}{\code{data.frame} Dataset with all observations for the test
set.}
}
}
\description{
Get train/test splits of the phenotypic MET dataset based on a
number of random k-folds partitions determined by the user, according to the
type CV2. Creation of the list of train/test splits based on phenotypic data,
so that all the Year x Location phenotypic observations from the phenotypic
MET dataset are assigned randomly to k-fold partitions (prediction of
incomplete field trials).
}
\references{
\insertRef{jarquin2017increasing}{learnMET}
}
\author{
Cathy C. Westhues \email{cathy.jubin@uni-goettingen.de}
}
