% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_trait_MET_cv.R
\name{predict_trait_MET_cv}
\alias{predict_trait_MET_cv}
\title{Cross-validation procedure for phenotypic prediction}
\usage{
predict_trait_MET_cv(
  METData,
  trait,
  method = "xgboost",
  use_selected_markers = T,
  geno_information = c("SNPs", "PCs", "PCs_G"),
  num_pcs = 200,
  lat_lon_included = T,
  year_included = F,
  cv_type = c("cv0", "cv1", "cv2"),
  cv0_type = c("leave-one-environment-out", "leave-one-site-out", "leave-one-year-out",
    "forward-prediction"),
  nb_folds_cv1 = 5,
  repeats_cv1 = 50,
  nb_folds_cv2 = 5,
  repeats_cv2 = 50,
  include_env_predictors = T,
  list_env_predictors = NULL,
  plot_PA = T,
  path_plot_PA = "",
  ...
)
}
\arguments{
\item{METData}{\code{list} An object created by the initial function of the
package create_METData().}

\item{trait}{\code{character} Name of the trait to predict.}

\item{method}{\code{character} specifying the predictive model to use.}

\item{use_selected_markers}{\code{Logical} Whether to use a subset of markers
obtained from a previous step (see function select_markers()).}

\item{geno_information}{\code{character} Method to use to incorporate
genomic information in predictions. T}
}
\value{
Object of class....
}
\description{
Implement trait prediction based on SNP and environmental data
with a range of prediction methods (Machine Learning)

This function should be used to assess the predictive ability according to
a cross-validation scheme determined by the user.
}
\author{
Cathy C. Jubin \email{cathy.jubin@uni-goettingen.de}
}
