% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_trait_MET_cv.R
\name{predict_trait_MET_cv}
\alias{predict_trait_MET_cv}
\title{Trait prediction based on SNP and environmental data
This function should be used to assess the predictive ability according to
various cross-validation schemes.}
\usage{
predict_trait_MET_cv(
  METData,
  trait,
  method = "xgboost",
  use_selected_markers = T,
  geno_information = c("SNPs", "PCs", "PCs_G"),
  num_pcs = 200,
  lat_lon_included = T,
  year_included = F,
  cv_type = c("cv0", "cv1", "cv2"),
  cv0_type = c("leave-one-environment-out", "leave-one-site-out", "leave-one-year-out",
    "forward-prediction"),
  nb_folds_cv1 = 5,
  repeats_cv1 = 50,
  nb_folds_cv2 = 5,
  repeats_cv2 = 50,
  include_env_predictors = T,
  list_env_predictors = NULL,
  ...
)
}
\arguments{
\item{use_selected_markers}{\code{Logical} Whether to use a subset of markers
obtained from a previous step (see function select_markers()).}

\item{METData.}{An object created by the initial function of the package,
"create_METData.R"}

\item{trait.}{\code{character} Name of the trait under study for which a}
}
\description{
Trait prediction based on SNP and environmental data
This function should be used to assess the predictive ability according to
various cross-validation schemes.
}
