% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_trait_MET_cv.R
\name{predict_trait_MET_cv}
\alias{predict_trait_MET_cv}
\title{Cross-validation procedure for phenotypic prediction of crop varieties.}
\usage{
predict_trait_MET_cv(
  METData,
  trait,
  prediction_method,
  use_selected_markers = F,
  build_haplotypes = F,
  list_selected_markers_manual = NULL,
  lat_lon_included = F,
  year_included = F,
  cv_type = "cv0",
  cv0_type = "leave-one-environment-out",
  nb_folds_cv1 = 5,
  repeats_cv1 = 50,
  nb_folds_cv2 = 5,
  repeats_cv2 = 50,
  include_env_predictors = T,
  list_env_predictors = NULL,
  seed = NULL,
  save_splits = F,
  save_processing = F,
  path_folder,
  compute_vip = F,
  ...
)
}
\arguments{
\item{METData}{\code{list} An object created by the initial function of the
package \code{\link[=create_METData]{create_METData()}}.}

\item{trait}{\code{character} Name of the trait to predict.}

\item{prediction_method}{\code{character} specifying the predictive model to use.
Options are currently \code{xgb_reg_1} (gradient boosted trees), \code{xgb_reg_2} ,
\code{xgb_reg_3}, \code{DL_reg_1} (multilayer perceptrons), \code{DL_reg_2}, \code{DL_reg_3},
\code{stacking_reg_1} (stacked models), \code{stacking_reg_2}, \code{stacking_reg_3}.}

\item{use_selected_markers}{A \code{Logical} indicating whether to use a
subset of markers  identified via single-environment GWAS or based on the
table of marker effects obtained via Elastic Net as predictor variables,
when main genetic effects are modeled with principal components. \cr
If \code{use_selected_markers} is \code{TRUE}, and if \code{list_selected_markers_manual}
is \code{NULL}, then the \code{\link[=select_markers]{select_markers()}} function will be called in the
pipeline.
\strong{For more details, see \code{\link[=select_markers]{select_markers()}}}}

\item{lat_lon_included}{\code{logical} indicates if longitude and latitude
data should be used as numeric predictors. Default is \code{FALSE}.}

\item{year_included}{\code{logical} indicates if year factor should be used
as predictor variable. Default is \code{FALSE}.}

\item{cv_type}{A \code{character} with one out of \code{cv0} (prediction of new
environments), \code{cv00} (prediction of new genotypes in new environments),
\code{cv1} (prediction of new genotypes) or \code{cv2} (prediction of incomplete
field trials). Default is \code{cv0}.}

\item{cv0_type}{A \code{character} with one out of
\code{leave-one-environment-out}, \code{leave-one-site-out},\code{leave-one-year-out},
\code{forward-prediction}. Default is \code{leave-one-environment-out}.}

\item{nb_folds_cv1}{A \code{numeric} Number of folds used in the CV1 scheme.
Default is 5.}

\item{repeats_cv1}{A \code{numeric} Number of repeats in the CV1 scheme.
Default is 50.}

\item{nb_folds_cv2}{A \code{numeric} Number of folds used in the CV2 scheme.
Default is 5.}

\item{repeats_cv2}{A \code{numeric} Number of repeats in the CV2 scheme.
Default is 50.}

\item{include_env_predictors}{A \code{logical} indicating whether
environmental covariates characterizing each environment should be used in
predictions.}

\item{list_env_predictors}{A \code{character} vector containing the names
of the environmental predictors which should be used in predictions.
\strong{By default \code{NULL}: all environmental predictors included in the
env_data table of the \code{METData} object will be used.}}

\item{seed}{\code{integer} Seed value. Default is \code{NULL}. By default, a
random seed will be generated.}

\item{save_splits}{A \code{Logical} to indicate if the train/test splits
should be saved.}

\item{save_processing}{a \code{logical} indicating whether the processing
steps obtained from the \code{\link[=get_splits_processed_with_method]{get_splits_processed_with_method()}} functions should be saved in a .RDS
object. Default is \code{FALSE}.}

\item{path_folder}{a \code{character} indicating the full path where the .RDS
object and plots generated during the analysis should be saved (do not use
a Slash after the name of the last folder). Default is \code{NULL}.}

\item{...}{Arguments passed to the \code{\link[=get_splits_processed_with_method]{get_splits_processed_with_method()}}
function.}
}
\value{
A \code{list} object of class \code{met_cv} with the following items:
\describe{
\item{list_results_cv}{\code{list} of \code{res_fitted_split} elements.
The length of this list corresponds to the number of training/test set
partitions.
Within each list, different sub-elements are provided as output and
listed below:
\describe{
\item{}
}\preformatted{Detailed prediction results for each split of the
data within each element of this list.
\}
\\item\{seed_used\}\{\code{integer} Seed used to generate the
cross-validation splits.\}
\\item\{cv_type\}\{\code{integer} Seed used to generate the
cross-validation splits.\}
\}
}
}
\description{
Implement trait prediction based on SNP and environmental data
with selection of prediction methods among Machine Learning approaches.

This function should be used to assess the predictive ability according to
a cross-validation scheme determined by the user.
}
\examples{
library (learnMET)

# Evaluation of the dataset with a CV0 cross-validation scenario with a
# stacked model using and a LASSO model as the meta-model. 
rescv0_1 <- predict_trait_MET_cv(
  METData = METdata_indica, 
  trait = 'PH', 
  method_processing = 'stacking_reg_1',
  use_selected_markers = F,
  num_pcs = 300,
  lat_lon_included = F,
  year_included = F,
  cv_type = 'cv0',
  cv0_type = 'leave-one-year-out',
  nb_folds_cv1 = 3,
  repeats_cv1 = 2,
  nb_folds_cv2 = 5,
  repeats_cv2 = 50,
  kernel_G = 'linear',
  include_env_predictors = T,
  list_env_predictors = NULL,
  save_processing  = T,
  seed = 100,
  path_folder = 'user1/myDocuments/predictions_indica_rice/elasticnet_FS/cv0'
  # A directory is created 
  )
  
  
}
\author{
Cathy C. Westhues \email{cathy.jubin@uni-goettingen.de}
}
