% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_trait.R
\name{predict_trait_MET_cv}
\alias{predict_trait_MET_cv}
\title{Trait prediction based on genetic and environmental data
This function assumes that METData$pheno contains all}
\usage{
predict_trait_MET_cv(
  METData,
  trait,
  method = "xgboost",
  use_selected_markers = T,
  lat_lon_included = T,
  year_included = F,
  cv_type = c("cv0", "cv1", "cv2"),
  cv0_type = c("leave-one-environment-out", "leave-one-site-out", "leave-one-year-out"),
  nb_folds_cv1 = 5,
  repeats_cv1 = 50,
  nb_folds_cv2 = 5,
  repeats_cv2 = 50,
  list_env_predictors = colnames(METData$env_data)[colnames(METData$env_data) \%notin\%
    c("IDenv", "year", "location", "longitude", "latitude")],
  ...
)
}
\arguments{
\item{METData.}{An object created by the initial function of the package,
"create_METData.R"}

\item{trait.}{\code{character} Name of the trait under study for which a}
}
\description{
Trait prediction based on genetic and environmental data
This function assumes that METData$pheno contains all
}
