% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_train_test_split.R
\name{processing_train_test_split}
\alias{processing_train_test_split}
\title{Processing and selecting predictors to fit the model}
\usage{
processing_train_test_split(
  split,
  list_env_predictors,
  include_env_predictors,
  lat_lon_included,
  year_included,
  use_selected_markers
)
}
\arguments{
\item{use_selected_markers}{}
}
\value{

}
\description{
Processing and selecting predictors to fit the model
}
