% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_train_test_split.R
\name{processing_train_test_split}
\alias{processing_train_test_split}
\title{Processing and selecting predictors to fit the model}
\usage{
processing_train_test_split(
  split,
  geno_data,
  geno_information,
  use_selected_markers,
  SNPs,
  list_env_predictors,
  include_env_predictors,
  lat_lon_included,
  year_included
)
}
\arguments{
\item{use_selected_markers}{}
}
\value{
a \code{list} object containing:
\itemize{
\item training: \code{data.frame} Training set
\item test: \code{data.frame} Test set
\item rec: \code{recipe} object with the different steps to implement
as processing steps on the training dataset. Same transformations applied
on the test set.}
}
}
\description{
Processing and selecting predictors to fit the model
}
\author{
Cathy C. Jubin \email{cathy.jubin@uni-goettingen.de}
}
