% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_train_test_split_kernel.R
\name{processing_train_test_split_kernel}
\alias{processing_train_test_split_kernel}
\title{Processing and selecting predictors to fit the model}
\usage{
processing_train_test_split_kernel(
  split,
  geno_data,
  geno_information,
  use_selected_markers,
  SNPs,
  list_env_predictors,
  include_env_predictors,
  lat_lon_included = F,
  year_included = F,
  ...
)
}
\arguments{
\item{geno}{threshold 0.95 correlation}
}
\value{
a \code{list} object containing:
\itemize{
\item training: \code{data.frame} Training set
\item test: \code{data.frame} Test set
\item rec_G: \code{recipe} object with the different steps to process molecular
marker dataset from the training set. Same transformations applied on the
test set.
\item rec_E: \code{recipe} object with the different steps to process
environmental data from the training set. Same transformations applied on
the test set.
\item rec_GE: \code{recipe} object with the different steps to process
the GxE dataset from the training set. Same transformations applied on
the test set.
}
}
\description{
Processing and selecting predictors to fit the model
}
\author{
Cathy C. Jubin \email{cathy.jubin@uni-goettingen.de}
}
