% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_raw_weather_data.R
\name{qc_raw_weather_data}
\alias{qc_raw_weather_data}
\title{Quality control on daily weather data}
\usage{
qc_raw_weather_data(daily_weather_data, info_environments, path_flagged_values)
}
\arguments{
\item{daily_weather_data}{a \code{data.frame} which contains the following
mandatory columns:
\enumerate{
\item longitude \code{numeric}
\item latitude \code{numeric}
\item year \code{numeric}
\item location \code{character}
\item YYYYMMDD \code{Date}
\item IDenv \code{character}
\item DOY \code{integer}
}
Available weather data provided by user must be a subset of the following
weather variable names.
Column names of weather variables must be given as following:
(\strong{For envs with raw daily weather data, T2M, T2M_MIN, T2M_MAX and
PRECTOTCORR are mandatory to provide and must be given without missing values,
which implies that any imputation step should be performed before providing
this dataset to the package.
.}):
\enumerate{
\item T2M \code{numeric} Daily mean temperature (째C)
\item T2M_MIN \code{numeric} Daily minimum temperature (째C)
\item T2M_MAX \code{numeric} Daily maximum temperature (째C)
\item PRECTOTCORR \code{numeric} Daily total precipitation (mm)
\item RH2M \code{numeric} Daily mean relative humidity (\%)
\item RH2M_MIN \code{numeric} Daily minimum relative humidity (\%)
\item RH2M_MAX \code{numeric} Daily maximum relative humidity (\%)
\item daily_solar_radiation \code{numeric} daily solar radiation
(MJ/m^2/day)
\item top_atmosphere_insolation \code{numeric} Top-of-atmosphere
Insolation (MJ/m^2/day)
\item T2MDEW \code{numeric} Dew Point (째C)
}}
}
\value{
a processed \code{data.frame} after quality check with the same
columns as before the QC. \cr
Vapor pressure deficit is calculated if T2M_MIN, T2M_MAX, and either
RH2M_MIN + RH2M_MAX  or only RH2M are provided.   \cr
\strong{
Warning messages are also thrown if some observations do not pass either
the range test, persistence test or the internal consistency test. A
data.frame with dubious values replaced by a "flagged" character is
provided, along with explanations why these data were flagged (column
"reason") is provided as output. None of the flagged values is assigned as
missing values or transformed; therefore we strongly recommend the user to
have a second look at the daily weather data provided and to correct
potential dubious values indicated by the output of the present function.}
\cr
\strong{
Solar radiation or wind data are automatically retrieved from NASA, if they
are not provided without any missing data by the user. As for any other
weather variable used in this function, these data cannot be only partially
provided (no missing values accepted).}
}
\description{
This function checks range of values for \code{METData} and implements
various test on daily weather data (persistence tests, internal
consistency tests) provided by the user.
}
\references{
\insertRef{zotarelli2010step}{learnMET}
}
\author{
Cathy C. Westhues \email{cathy.jubin@uni-goettingen.de}
}
