% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_raw_weather_data.R
\name{qc_raw_weather_data}
\alias{qc_raw_weather_data}
\title{Quality control on daily weather data}
\usage{
qc_raw_weather_data(daily_weather_data, info_environments, path_flagged_values)
}
\arguments{
\item{daily_weather_data}{a \code{data.frame} which contains the following
mandatory columns:
\enumerate{
\item longitude \code{numeric}
\item latitude \code{numeric}
\item year \code{numeric}
\item location \code{character}
\item YYYYMMDD \code{Date} Date of the daily observation
\item IDenv \code{character} Environmt ID written Location_Year
\item T2M \code{numeric} Average mean temperature (degree Celsius)
\item T2M_MIN \code{numeric} Min. temperature (degree Celsius)
\item T2M_MAX \code{numeric} Max. temperature (degree Celsius)
\item PRECTOTCORR \code{numeric} Total daily precipitation (mm)
}
Additional weather data provided by user must be a subset of the following
weather variable names (= next columns):
(\strong{Any imputation step should be performed before providing
this daily weather dataset to the package. }):
\enumerate{
\item RH2M \code{numeric} Daily mean relative humidity (\%)
\item RH2M_MIN \code{numeric} Daily minimum relative humidity (\%)
\item RH2M_MAX \code{numeric} Daily maximum relative humidity (\%)
\item daily_solar_radiation \code{numeric} daily solar radiation
(MJ/m^2/day)
\item T2MDEW \code{numeric} Dew Point (°C)
}
Default is \code{NULL}.}
}
\value{
a processed \code{data.frame} after quality check with the same
columns as before the QC. \cr
Vapor pressure deficit is calculated if T2M_MIN, T2M_MAX, and either
RH2M_MIN + RH2M_MAX  or only RH2M are provided.   \cr
\strong{
Warning messages are also thrown if some observations do not pass either
the range test, persistence test or the internal consistency test. A
data.frame, with dubious values signaled by a column flagged and with the
corresponding explanation in the column "reason", is provided as output.
None of the flagged values is assigned as missing values or transformed;
therefore we strongly recommend the user to have a second look at the daily
weather data provided and to correct potential dubious values indicated by
the output of the present function.}
\cr
\strong{
Solar radiation or wind data are automatically retrieved from NASA, if they
are not provided without any missing data by the user. As for any other
weather variable used in this function, these data cannot be only partially
provided (no missing values accepted).}
}
\description{
This function checks range of values for \code{METData} and implements
various test on daily weather data (persistence tests, internal
consistency tests) provided by the user.
}
\references{
\insertRef{zotarelli2010step}{learnMET}
}
\author{
Cathy C. Westhues \email{cathy.jubin@uni-goettingen.de}
}
