% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_stacking_reg.R
\name{svm_stacking_reg}
\alias{svm_stacking_reg}
\alias{new_svm_stacking_reg}
\alias{validate_svm_stacking_reg}
\title{}
\usage{
new_svm_stacking_reg(
  split,
  trait,
  geno_data,
  env_predictors,
  info_environments,
  unique_EC_by_geno,
  geno_information,
  use_selected_markers,
  SNPs,
  list_env_predictors,
  include_env_predictors,
  lat_lon_included,
  year_included
)

svm_stacking_reg(
  split,
  trait,
  geno_data,
  env_predictors,
  info_environments,
  unique_EC_by_geno,
  geno_information,
  use_selected_markers,
  SNPs,
  list_env_predictors,
  include_env_predictors,
  lat_lon_included,
  year_included
)

validate_svm_stacking_reg(x)
}
\description{
blabla
}
