% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_importance.R
\name{variable_importance_split}
\alias{variable_importance_split}
\alias{variable_importance_split.default}
\alias{variable_importance_split.DL_reg}
\alias{variable_importance_split.xgb_reg}
\alias{variable_importance_split.stacking_reg_1}
\title{S3 method used to fit an object of class \code{xgb_reg}, \code{DL_reg},
\code{stacking_reg_1}, \code{stacking_reg_2} or \code{stacking_reg_3}.}
\usage{
variable_importance_split(object, ...)

\method{variable_importance_split}{default}(x, ...)

\method{variable_importance_split}{DL_reg}(fitting_all_splits, trait)

\method{variable_importance_split}{xgb_reg}(fitting_all_splits, trait)

\method{variable_importance_split}{stacking_reg_1}(fitting_all_splits, trait)
}
\arguments{
\item{object}{an object of class \code{xgb_reg}, \code{DL_reg},
\code{stacking_reg_1}, \code{stacking_reg_2} or \code{stacking_reg_3}.}
}
\description{
S3 dispatching method for objects of class \code{xgb_reg}, \code{DL_reg},
\code{stacking_reg_1}, \code{stacking_reg_2} or \code{stacking_reg_3}.
}
\author{
Cathy C. Westhues \email{cathy.jubin@uni-goettingen.de}
}
