% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_importance.R
\name{variable_importance_split}
\alias{variable_importance_split}
\alias{variable_importance_split.default}
\alias{variable_importance_split.fitted_DL_reg_1}
\alias{variable_importance_split.fitted_DL_reg_2}
\alias{variable_importance_split.fitted_xgb_reg_1}
\alias{variable_importance_split.fitted_xgb_reg_2}
\alias{variable_importance_split.fitted_stacking_reg_1}
\alias{variable_importance_split.fitted_stacking_reg_2}
\alias{variable_importance_split.fitted_stacking_reg_3}
\alias{variable_importance_split.fitted_rf_reg_1}
\alias{variable_importance_split.fitted_rf_reg_2}
\title{Compute variable importance according to the machine learning algorithm used}
\usage{
variable_importance_split(object, ...)

\method{variable_importance_split}{default}(object, ...)

\method{variable_importance_split}{fitted_DL_reg_1}(object)

\method{variable_importance_split}{fitted_DL_reg_2}(object)

\method{variable_importance_split}{fitted_xgb_reg_1}(
  object,
  path_plot,
  type = "model_specific",
  permutations = 10,
  unseen_data = F
)

\method{variable_importance_split}{fitted_xgb_reg_2}(object)

\method{variable_importance_split}{fitted_stacking_reg_1}(object)

\method{variable_importance_split}{fitted_stacking_reg_2}(object)

\method{variable_importance_split}{fitted_stacking_reg_3}(object)

\method{variable_importance_split}{fitted_rf_reg_1}(object)

\method{variable_importance_split}{fitted_rf_reg_2}(object)

\method{variable_importance_split}{fitted_rf_reg_2}(object)

\method{variable_importance_split}{fitted_rf_reg_1}(object)
}
\arguments{
\item{object}{an object of class \code{res_object}}

\item{type}{\code{model_specific} or \code{model_agnostic}}

\item{permutations}{By default, equal to 10.}

\item{unseen_data}{}
}
\description{
Variable importance can be calculated based on model-specific and
model-agnostic approaches
}
\references{
\insertRef{breiman2001random}{learnMET}
\insertRef{molnar2022}{learnMET}
}
\author{
Cathy C. Westhues \email{cathy.jubin@hotmail.com}
}
